% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_admid}
\alias{add_admid}
\title{add_admid}
\usage{
add_admid(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(model : Model) Pharmpy model
}
\description{
Add an admid column to the model dataset and datainfo. Dependent on the
presence of a CMT column in order to add admid correctly.

When generated, admids of events in between doses is set to the last used
admid.
}
\seealso{
get_admid : Get or create an admid column

get_cmt : Get or create a cmt column
}
