% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_administration}
\alias{add_administration}
\title{add_administration}
\usage{
add_administration(td, arm, variable, amount, time_points, start_time = 0)
}
\arguments{
\item{td}{(TrialDesign) TrialDesign to add to}

\item{arm}{(str) Name of the arm}

\item{variable}{(str) Name of the dose variable (e.g. AMT)}

\item{amount}{(numeric) The dose amount}

\item{time_points}{(array(numeric)) List of the time points for the administration}

\item{start_time}{(numeric) Set a start_time to offset the time_points. Default 0.0}
}
\value{
(TrialDesign) An updated TrialDesign
}
\description{
Add observations to an arm in a trial design
}
\examples{
\dontrun{
td <- create_trial_design()
td <- add_arm(td, name="Drug", size=100)
td <- add_observations(td, arm="Drug", variable="DV", time_points=[0.0, 1.0, 2.0, 4.0, 16.0])
td <- add_administration(td, arm="Drug", variable="AMT", amount=10.0, time_points=[0.0, 8.0])
td
}

}
