% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_arm}
\alias{add_arm}
\title{add_arm}
\usage{
add_arm(td, name, size)
}
\arguments{
\item{td}{(TrialDesign) TrialDesign to add to}

\item{name}{(str) Name of the arm}

\item{size}{(numeric) Size of the arm (number of subjects)}
}
\value{
(TrialDesign) An updated TrialDesign
}
\description{
Add an empty arm to a trial design
}
\examples{
\dontrun{
td <- create_trial_design()
td <- add_arm(td, name="Placebo", size=100)
td
}

}
