% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_cmt}
\alias{add_cmt}
\title{add_cmt}
\usage{
add_cmt(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(model : Model) Pharmpy model
}
\description{
Add a CMT column to the model dataset and datainfo if not existed

In case of multiple doses, this method is dependent on the presence of an
admid column to correctly number each dose.

NOTE : Existing CMT is based on datainfo type being set to 'compartment'
and a column named 'CMT' can be replaced
}
\seealso{
get_admid : Get or create an admid column

get_cmt : Get or create a cmt column
}
