% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_estimation_step}
\alias{add_estimation_step}
\title{add_estimation_step}
\usage{
add_estimation_step(
  model,
  method,
  interaction = TRUE,
  options = NULL,
  idx = NULL
)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{method}{(str) estimation method to change to}

\item{interaction}{(logical) whether to use interaction or not, default is true}

\item{options}{(list) any additional options. Note that this removes old options}

\item{idx}{(int) index of estimation step, default is NULL (adds step last)}
}
\value{
model (Model)
}
\description{
Add estimation step

Adds estimation step for a model in a given index. Methods currently supported are:
FO, FOCE, ITS, LAPLACE, IMPMAP, IMP, SAEM
}
