% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_iiv}
\alias{add_iiv}
\title{add_iiv}
\usage{
add_iiv(
  model,
  list_of_parameters,
  expression,
  operation = "*",
  eta_names = NULL
)
}
\arguments{
\item{model}{(Model) Pharmpy model to add new IIVs to.}

\item{list_of_parameters}{(str, vector) Name/names of parameter to add new IIVs to.}

\item{expression}{(str, vector) Effect/effects on eta. Either abbreviated (see above) or custom.}

\item{operation}{(str, vector, optional) Whether the new IIV should be added or multiplied (default).}

\item{eta_names}{(str, vector, optional) Custom name/names of new eta

Model
Reference to the same model}
}
\description{
Adds IIVs to :class:\code{pharmpy.model}.

Effects that currently have templates are:
\itemize{
\item Additive (\emph{add})
\item Proportional (\emph{prop})
\item Exponential (\emph{exp})
\item Logit (\emph{logit})
}

For all except exponential the operation input is not needed. Otherwise user specified
input is supported. Initial estimates for new etas are 0.09.
}
\seealso{
add_iov

remove_iiv

remove_iov
}
