% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_indirect_effect}
\alias{add_indirect_effect}
\title{add_indirect_effect}
\usage{
add_indirect_effect(model, expr, prod = TRUE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{expr}{(str) Production (TRUE) (default) or degradation (FALSE)}

\item{prod}{(logical) Name of PD effect function. Valid names are: linear, emax, sigmoid and step}
}
\value{
(Model) Pharmpy model object
}
\description{
Add indirect (turnover) effect

The concentration (equation could not be rendered, see API doc on website)
\itemize{
\item Production:
(equation could not be rendered, see API doc on website)
\item Degradation:
(equation could not be rendered, see API doc on website)
}

(equation could not be rendered, see API doc on website)
Baseline (equation could not be rendered, see API doc on website)

Models:
\itemize{
\item Linear:
(equation could not be rendered, see API doc on website)
\item Emax:
(equation could not be rendered, see API doc on website)
\item Sigmoidal:
(equation could not be rendered, see API doc on website)
}
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- add_indirect_effect(model, expr='linear', prod=TRUE)
}

}
