% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_individual_parameter}
\alias{add_individual_parameter}
\title{add_individual_parameter}
\usage{
add_individual_parameter(model, name)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{name}{(str) Name of individual/pk parameter}
}
\description{
Add an individual or pk parameter to a model
}
