\name{add_lag_time}
\alias{add_lag_time}
\title{add_lag_time}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.add_lag_time}{Link to Python API reference} (e.g. for correct rendering of equations)

Add lag time to the dose compartment of model. Initial estimate for lag time is set the

previous lag time if available, otherwise it is set to the time of first observation/2 is

used.
}
\usage{
add_lag_time(model)
}
\arguments{

}
\value{

}
