% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_metabolite}
\alias{add_metabolite}
\title{add_metabolite}
\usage{
add_metabolite(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Pharmpy model object
}
\description{
Adds a metabolite compartment to a model

The flow from the central compartment to the metabolite compartment
will be unidirectional.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- add_metabolite(model)
}

}
