% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_observations}
\alias{add_observations}
\title{add_observations}
\usage{
add_observations(td, arm, variable, time_points, start_time = 0)
}
\arguments{
\item{td}{(TrialDesign) TrialDesign to add to}

\item{arm}{(str) Name of the arm}

\item{variable}{(str) Name of the variable to observe}

\item{time_points}{(array(numeric)) List of the time points for the observations}

\item{start_time}{(numeric) Set a start_time to offset the time_points. Default 0.0}
}
\value{
(TrialDesign) An updated TrialDesign
}
\description{
Add observations to an arm in a trial design
}
\examples{
\dontrun{
td <- create_trial_design()
td <- add_arm(td, name="Placebo", size=100)
td <- add_observations(td, arm="Placebo", variable="DV", time_points=[0.0, 1.0, 2.0, 4.0, 16.0])
td
}

}
