% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_output_variables}
\alias{add_output_variables}
\title{add_output_variables}
\usage{
add_output_variables(model, variables, append = TRUE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{variables}{(array(str)) List of variables to add}

\item{append}{(logical) Set to false to overwrite all variables}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Add output variables to an execution step
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$execution_steps[-1].variables
model <- add_output_variables(model, c('CL'))
model$execution_steps[-1].variables
}
}
\seealso{
add_predictions

add_residuals

remove_predictions

remove_residuals
}
