\name{add_parameter}
\alias{add_parameter}
\title{add_parameter}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.add_parameter}{Link to Python API reference} (e.g. for correct rendering of equations)

Add an individual or pk parameter to a model
}
\usage{
add_parameter(model, name)
}
\arguments{

}
\value{

}
