% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_parameter_uncertainty_step}
\alias{add_parameter_uncertainty_step}
\title{add_parameter_uncertainty_step}
\usage{
add_parameter_uncertainty_step(model, parameter_uncertainty_method)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameter_uncertainty_method}{(str) Parameter uncertainty method to use}
}
\value{
(Model) Pharmpy model object
}
\description{
Adds parameter uncertainty step to the final estimation step
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_estimation_step(model, 'FOCE', parameter_uncertainty_method=NULL)
model <- add_parameter_uncertainty_step(model, 'SANDWICH')
ests <- model$estimation_steps
ests[1]
}
}
\seealso{
add_estimation_step

set_estimation_step

remove_estimation_step

append_estimation_step_options

remove_parameter_uncertainty_step

set_evaluation_step
}
