% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_placebo_model}
\alias{add_placebo_model}
\title{add_placebo_model}
\usage{
add_placebo_model(model, expr, operator = "*")
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{expr}{(str) Name of placebo/disease progression effect function.}

\item{operator}{(str) Operator to use for combining the baseline with the placebo/disease progression}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Add a placebo or disease progression effect to a model.

warning:: This function is under development.
\itemize{
\item linear
}

(equation could not be rendered, see API doc on website)
\itemize{
\item exp_decrease
}

(equation could not be rendered, see API doc on website)
\itemize{
\item exp_increase
}

(equation could not be rendered, see API doc on website)
\itemize{
\item tmax
}

(equation could not be rendered, see API doc on website)

(equation could not be rendered, see API doc on website)
}
\examples{
\dontrun{
model <- create_basic_pd_model()
model <- add_placebo_model(model, "linear")
model$statements$find_assignment("PDP")
}

}
