% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_predictions}
\alias{add_predictions}
\title{add_predictions}
\usage{
add_predictions(model, pred)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{pred}{(array(str)) List of predictions (e.g. c('IPRED', 'PRED'))}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Add predictions and/or residuals

Add predictions to estimation step.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$execution_steps[-1].predictions
model <- add_predictions(model, c('IPRED'))
model$execution_steps[-1].predictions
}
}
\seealso{
remove_predictions

remove_residuals

set_estimation_step

add_estimation_step

remove_estimation_step

append_estimation_step_options

add_parameter_uncertainty_step

remove_parameter_uncertainty_step
}
