% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_time_after_dose}
\alias{add_time_after_dose}
\title{add_time_after_dose}
\usage{
add_time_after_dose(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Pharmpy model object
}
\description{
Calculate and add a TAD column to the dataset
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- add_time_after_dose(model)
}
}
\seealso{
add_time_of_last_dose : Add time of last dose to model
}
