% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_time_of_last_dose}
\alias{add_time_of_last_dose}
\title{add_time_of_last_dose}
\usage{
add_time_of_last_dose(model, name = "TDOSE")
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{name}{(str) Name of time of last dose variable}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Add a variable for time of last dose to the model
}
\seealso{
add_time_after_dose : Add time after dose to dataset
}
