\name{additive_error}
\alias{additive_error}
\title{additive_error}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.additive_error}{Link to Python API reference} (e.g. for correct rendering of equations)

Set an additive error model. Initial estimate for new sigma is :math:`(min(DV)/2)²`.
}
\usage{
additive_error(model)
}
\arguments{

}
\value{

}
