\name{bolus_absorption}
\alias{bolus_absorption}
\title{bolus_absorption}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.bolus_absorption}{Link to Python API reference} (e.g. for correct rendering of equations)

Set or change to bolus absorption rate.
}
\usage{
bolus_absorption(model)
}
\arguments{

}
\value{

}
