% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{broadcast_log}
\alias{broadcast_log}
\title{broadcast_log}
\usage{
broadcast_log(context, broadcaster = NULL)
}
\arguments{
\item{context}{(Context) Broadcast the log of this context}

\item{broadcaster}{(str (optional)) Name of the broadcaster to use. Default is to use the same as was original used.}
}
\description{
Broadcast the log of a context

Default is to use the same broadcaster, but optionally another broadcaster could
be used.
}
