% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{bump_model_number}
\alias{bump_model_number}
\title{bump_model_number}
\usage{
bump_model_number(model, path = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{path}{(str (optional)) Default is to not look for files.}
}
\value{
(Model) Updated Pharmpy model
}
\description{
If the model name ends in a number increase it

If path is set increase the number until no file exists
with the same name in path.
If model name does not end in a number do nothing.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- model$replace(name="run2")
model <- bump_model_number(model)
model$name
}

}
