% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_corr_from_prec}
\alias{calculate_corr_from_prec}
\title{calculate_corr_from_prec}
\usage{
calculate_corr_from_prec(precision_matrix)
}
\arguments{
\item{precision_matrix}{(data.frame) Precision matrix}
}
\value{
(data.frame) Correlation matrix
}
\description{
Calculate correlation matrix from a precision matrix
}
\examples{
\dontrun{
results <- load_example_modelfit_results("pheno")
prec <- results$precision_matrix
prec
calculate_corr_from_prec(prec)
}
}
\seealso{
calculate_se_from_cov : Standard errors from covariance matrix

calculate_se_from_prec : Standard errors from precision matrix

calculate_corr_from_cov : Correlation matrix from covariance matrix

calculate_cov_from_prec : Covariance matrix from precision matrix

calculate_cov_from_corrse : Covariance matrix from correlation matrix and standard errors

calculate_prec_from_cov : Precision matrix from covariance matrix

calculate_prec_from_corrse : Precision matrix from correlation matrix and standard errors
}
