% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_eta_shrinkage}
\alias{calculate_eta_shrinkage}
\title{calculate_eta_shrinkage}
\usage{
calculate_eta_shrinkage(model, sd = FALSE)
}
\description{
Calculate eta shrinkage for each eta

Variance = FALSE to get sd scale
}
