% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_parameters_from_ucp}
\alias{calculate_parameters_from_ucp}
\title{calculate_parameters_from_ucp}
\usage{
calculate_parameters_from_ucp(model, scale, ucps)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{scale}{(UCPSCale) A parameter scale}

\item{ucps}{(data.frame or list) Series of parameter values}
}
\value{
(data.frame) Parameters on the normal scale
}
\description{
Scale parameter values from ucp to normal scale
}
\examples{
\dontrun{
model <- load_example_model("pheno")
scale <- calculate_ucp_scale(model)
values <- {'THETA(1)': 0.1, 'THETA(2)': 0.1, 'THETA(3)': 0.1, 'OMEGA(1,1)': 0.1, 'OMEGA(2,2)': 0.1, 'SIGMA(1,1)': 0.1}
calculate_parameters_from_ucp(model, scale, values)
}
}
\seealso{
calculate_ucp_scale : Calculate the scale for conversion from ucps
}
