% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_se_from_cov}
\alias{calculate_se_from_cov}
\title{calculate_se_from_cov}
\usage{
calculate_se_from_cov(cov)
}
\arguments{
\item{cov}{(data.frame) Input covariance matrix}
}
\value{
(data.frame) Standard errors
}
\description{
Calculate standard errors from a covariance matrix
}
\examples{
\dontrun{
results <- load_example_modelfit_results("pheno")
cov <- results$covariance_matrix
cov
calculate_se_from_cov(cov)
}
}
\seealso{
calculate_se_from_prec : Standard errors from precision matrix

calculate_corr_from_cov : Correlation matrix from covariance matrix

calculate_cov_from_prec : Covariance matrix from precision matrix

calculate_cov_from_corrse : Covariance matrix from correlation matrix and standard errors

calculate_prec_from_cov : Precision matrix from covariance matrix

calculate_prec_from_corrse : Precision matrix from correlation matrix and standard errors

calculate_corr_from_prec : Correlation matrix from precision matrix
}
