% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_se_from_inf}
\alias{calculate_se_from_inf}
\title{calculate_se_from_inf}
\usage{
calculate_se_from_inf(information_matrix)
}
\arguments{
\item{information_matrix}{(data.frame) Input information matrix}
}
\value{
(data.frame) Standard errors
}
\description{
Calculate standard errors from an information matrix
}
\examples{
\dontrun{
model <- load_example_model("pheno")
inf <- model$modelfit_results$information_matrix
inf
calculate_se_from_inf(inf)
}
}
\seealso{
calculate_se_from_cov : Standard errors from covariance matrix

calculate_corr_from_cov : Correlation matrix from covariance matrix

calculate_cov_from_inf : Covariance matrix from information matrix

calculate_cov_from_corrse : Covariance matrix from correlation matrix and standard errors

calculate_inf_from_cov : Information matrix from covariance matrix

calculate_inf_from_corrse : Information matrix from correlation matrix and standard errors

calculate_corr_from_inf : Correlation matrix from information matrix
}
