% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_ucp_scale}
\alias{calculate_ucp_scale}
\title{calculate_ucp_scale}
\usage{
calculate_ucp_scale(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(UCPScale) A scale object
}
\description{
Calculate a scale for unconstrained parameters for a model

The UCPScale object can be used to calculate unconstrained parameters
back into the normal parameter space.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
scale <- calculate_ucp_scale(model)
}
}
\seealso{
calculate_parameters_from_ucp : Calculate parameters from ucp:s
}
