% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{check_dataset}
\alias{check_dataset}
\title{check_dataset}
\usage{
check_dataset(model, dataframe = FALSE, verbose = FALSE)
}
\arguments{
\item{model}{(Model) Pharmpy model object}

\item{dataframe}{(logical) TRUE to return a DataFrame instead of printing to the console}

\item{verbose}{(logical) Print out all rules checked if TRUE else print only failed rules}
}
\value{
(data.frame) Only returns a DataFrame is dataframe=TRUE
}
\description{
Check dataset for consistency across a set of rules
}
