% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_setup.R
\name{check_setup}
\alias{check_setup}
\title{Checks setup of Pharmpy/pharmr}
\usage{
check_setup()
}
\description{
Checks if everything is setup correctly. The following things are checked:
\itemize{
\item If Python is installed and has correct version
\item If Pharmpy is available
\item If Pharmpy and pharmr version matches
}
}
