% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{cholesky_decompose}
\alias{cholesky_decompose}
\title{cholesky_decompose}
\usage{
cholesky_decompose(model, rvs = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{rvs}{(array(str) (optional)) Names of random variables to decompose. NULL means all etas and is the default}
}
\value{
(Model) An updated Pharmpy model
}
\description{
Cholesky decomposition of joint normally distributed random variables
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- create_joint_distribution(model, c('ETA_CL', 'ETA_VC'))
model <- cholesky_decompose(model)
model$statements
}

}
