\name{combined_error}
\alias{combined_error}
\title{combined_error}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.combined_error}{Link to Python API reference} (e.g. for correct rendering of equations)

Set a combined error model. Initial estimates for new sigmas are :math:`(min(DV)/2)²` for

proportional and 0.09 for additive.
}
\usage{
combined_error(model)
}
\arguments{

}
\value{

}
