\name{copy_model}
\alias{copy_model}
\title{copy_model}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.copy_model}{Link to Python API reference} (e.g. for correct rendering of equations)


}
\usage{
copy_model(model)
}
\arguments{

}
\value{

}
