% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{create_basic_kpd_model}
\alias{create_basic_kpd_model}
\title{create_basic_kpd_model}
\usage{
create_basic_kpd_model(dataset_path = NULL, driver = "ir")
}
\arguments{
\item{dataset_path}{(str (optional)) Optional path to a dataset}

\item{driver}{(str) Driver variable of the KPD model. Can either be 'ir' (virtual infusion
rate) or 'amount'.}
}
\value{
(Model) Pharmpy model
}
\description{
Creates a basic kpd model. The model will be a one compartment model.
delay. The elimination rate will be (equation could not be rendered, see API doc on website)
}
\examples{
\dontrun{
model <- create_basic_kpd_model()
}

}
