% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{create_basic_pd_model}
\alias{create_basic_pd_model}
\title{create_basic_pd_model}
\usage{
create_basic_pd_model(dataset_path = NULL)
}
\arguments{
\item{dataset_path}{(str (optional)) Optional path to a dataset}
}
\value{
(Model) Pharmpy model
}
\description{
Create a basic pd model.

The model will be very simple describing only a constant baseline effect
with an additive error.
}
\examples{
\dontrun{
model <- create_basic_pd_model()
}

}
