% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{create_config_template}
\alias{create_config_template}
\title{create_config_template}
\usage{
create_config_template()
}
\description{
Create a basic config file template

If a configuration file already exists it will not be overwritten
}
\examples{
\dontrun{
create_config_template()
}

}
