% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{create_context}
\alias{create_context}
\title{create_context}
\usage{
create_context(name, path = NULL)
}
\arguments{
\item{name}{(str) Name of the context}

\item{path}{(str (optional)) Path to where to put the context}
}
\description{
Create a new context

Currently a local filesystem context (i.e. a directory)
}
\examples{
\dontrun{
ctx <- create_context("myproject")
}

}
