% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{create_datainfo}
\alias{create_datainfo}
\title{create_datainfo}
\usage{
create_datainfo(path_or_df)
}
\arguments{
\item{path_or_df}{(str or data.frame) A path to a dataset or a dataset}
}
\value{
(DataInfo) DataInfo object
}
\description{
Create a DataInfo

Will assume NONMEM names of columns
}
