% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{create_dataset_from_design}
\alias{create_dataset_from_design}
\title{create_dataset_from_design}
\usage{
create_dataset_from_design(td)
}
\arguments{
\item{td}{(TrialDesign) TrialDesign to add to}
}
\value{
(data.frame) A dataset
}
\description{
Create a dataset matching a design
}
\examples{
\dontrun{
td <- create_trial_design()
td <- add_arm(td, name="Drug", size=100)
td <- add_observations(td, arm="Drug", variable="DV", time_points=[0.0, 1.0, 2.0, 4.0, 16.0])
td <- add_administration(td, arm="Drug", variable="AMT", amount=10.0, time_points=[0.0, 8.0])
df <- create_dataset_from_design(td)
df
}

}
