% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{create_report}
\alias{create_report}
\title{create_report}
\usage{
create_report(results, path)
}
\arguments{
\item{results}{(Results) Results for which to create report}

\item{path}{(Path) Path to report file}
}
\description{
Create standard report for results

The report will be an html created at specified path.
}
