\name{create_results}
\alias{create_results}
\title{create_results}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.create_results}{Link to Python API reference} (e.g. for correct rendering of equations)


}
\usage{
create_results(path, **kwargs)
}
\arguments{

}
\value{

}
