% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{create_rng}
\alias{create_rng}
\title{create_rng}
\usage{
create_rng(seed = NULL)
}
\arguments{
\item{seed}{(numeric (optional)) Seed for the random number generator or NULL (default) for a randomized seed. If seed
is generator it will be passed through.}
}
\value{
(Generator) Initialized numpy random number generator object
}
\description{
Create a new random number generator

Pharmpy functions that use random sampling take a random number generator or seed as input.
This function can be used to create a default new random number generator.
}
\examples{
\dontrun{
rng <- create_rng(23)
rng$standard_normal()
}

}
