% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{create_trial_design}
\alias{create_trial_design}
\title{create_trial_design}
\usage{
create_trial_design(idv_name = "TIME")
}
\arguments{
\item{idv_name}{(str) Optional name of the independent variable. Default is TIME}
}
\value{
(TrialDesign) An empty trial design
}
\description{
Create an empty trial design
}
\examples{
\dontrun{
td <- create_trial_design()
td$is_empty
td
}

}
