% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{display_odes}
\alias{display_odes}
\title{display_odes}
\usage{
display_odes(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(ODEDisplayer) A displayable object
}
\description{
Displays the ordinary differential equation system
}
\examples{
\dontrun{
model <- load_example_model("pheno")
display_odes(model)
}

}
