% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{drop_columns}
\alias{drop_columns}
\title{drop_columns}
\usage{
drop_columns(model, column_names, mark = FALSE)
}
\arguments{
\item{model}{(Model) Pharmpy model object}

\item{column_names}{(str or array(str)) List of column names or one column name to drop or mark as dropped}

\item{mark}{(logical) Default is to remove column from dataset. Set this to TRUE to only mark as dropped}
}
\value{
(Model) Reference to same model object
}
\description{
Drop columns from the dataset or mark as dropped
}
\examples{
\dontrun{
model <- load_example_model("pheno")
drop_columns(model, c('WGT', 'APGR'))
vector(model$dataset$columns)
}
}
\seealso{
drop_dropped_columns : Drop all columns marked as drop

undrop_columns : Undrop columns of model
}
