% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{drop_dropped_columns}
\alias{drop_dropped_columns}
\title{drop_dropped_columns}
\usage{
drop_dropped_columns(model)
}
\arguments{
\item{model}{(Model) Pharmpy model object}
}
\value{
(Model) Pharmpy model object
}
\description{
Drop columns marked as dropped from the dataset

NM-TRAN date columns will not be dropped by this function
even if marked as dropped.
Columns not specified in the datainfo ($INPUT for NONMEM)
will also be dropped from the dataset.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- drop_dropped_columns(model)
vector(model$dataset$columns)
}
}
\seealso{
drop_columns : Drop specific columns or mark them as drop
}
