% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{evaluate_eta_gradient}
\alias{evaluate_eta_gradient}
\title{evaluate_eta_gradient}
\usage{
evaluate_eta_gradient(model, etas = NULL, parameters = NULL, dataset = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{etas}{(data.frame (optional)) Optional list of eta values}

\item{parameters}{(list(str=numeric) (optional)) Optional list of parameters and values}

\item{dataset}{(data.frame (optional)) Optional dataset}
}
\value{
(data.frame) Gradient
}
\description{
Evaluate the numeric eta gradient

The gradient is evaluated at the current model parameter values
or optionally at the given parameter values.
The gradient is done for each data record in the model dataset
or optionally using the dataset argument.
The gradient is done at the current eta values
or optionally at the given eta values.

This function currently only support models without ODE systems
}
\examples{
\dontrun{
model <- load_example_model("pheno_linear")
etas <- model$modelfit_results$individual_estimates
evaluate_eta_gradient(model, etas=etas)
}
}
\seealso{
evaluate_epsilon_gradient : Evaluate the epsilon gradient
}
