% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{evaluate_expression}
\alias{evaluate_expression}
\title{evaluate_expression}
\usage{
evaluate_expression(model, expression)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{expression}{(str or sympy expression) Expression to evaluate

pd.Series
A series of one evaluated value for each data record}
}
\description{
Evaluate expression using model

Calculate the value of expression for each data record.
The expression can contain dataset columns, variables in model and
population parameters. If the model has parameter estimates these
will be used. Initial estimates will be used for non-estimated parameters.
}
