% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{explicit_odes}
\alias{explicit_odes}
\title{explicit_odes}
\usage{
explicit_odes(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Convert model from compartmental system to explicit ODE system
or do nothing if it already has an explicit ODE system
}
