% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{export_model_files}
\alias{export_model_files}
\title{export_model_files}
\usage{
export_model_files(context, destination_path = NULL, force = FALSE)
}
\arguments{
\item{context}{(Context) The context}

\item{destination_path}{(str (optional)) Path to export model files to, NULL means current working directory}

\item{force}{(logical) Allow file overwrite (default is FALSE)}
}
\description{
Exports all model files to specified directory.

Will export all model files generated/related to the external software used. Files will be named with model
name (from context) and original suffix (e.g. model.ctl for modelsearch_run1 -> modelsearch_run1.ctl). If no
suffix, file will be named model name and original name (e.g. mytab for modelsearch_run1 ->
modelsearch_run1_mytab)
}
\examples{
\dontrun{
ctx <- open_context("myrun")
export_model_files(ctx)
}

}
