% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{filter_dataset}
\alias{filter_dataset}
\title{filter_dataset}
\usage{
filter_dataset(model, expr)
}
\arguments{
\item{model}{(Model) Pharmpy model object}

\item{expr}{(str) expression for dataset query}
}
\value{
(Model) Pharmpy model object
}
\description{
Filter dataset according to expr and return a model with the filtered dataset.

Example: "DVID == 1" will filter the dataset so that only the rows with DVID = 1 remain.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$dataset
model <- filter_dataset(model, 'WGT < 1.4')
model$dataset
}

}
