% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{find_clearance_parameters}
\alias{find_clearance_parameters}
\title{find_clearance_parameters}
\usage{
find_clearance_parameters(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(vector) A vector of clearance parameters
}
\description{
Find clearance parameters in model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
find_clearance_parameters(model)
}

}
