% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{find_volume_parameters}
\alias{find_volume_parameters}
\title{find_volume_parameters}
\usage{
find_volume_parameters(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(vector) A vector of volume parameters
}
\description{
Find volume parameters in model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
find_volume_parameters(model)
}

}
