\name{first_order_absorption}
\alias{first_order_absorption}
\title{first_order_absorption}
\description{
Set or change to first order absorption rate.

\bold{Usage}

\code{first_order_absorption(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.first_order_absorption}{API reference} (for correct rendering of equations, tables etc.).
Set or change to first order absorption rate. Initial estimate for absorption rate is set

the previous rate if available, otherwise it is set to the time of first observation/2 is used.

This documentation was automatically generated from Pharmpy (0.22.10).}
\arguments{

\item{
model
}{
Model. Model to set or change to bolus absorption rate
}
}
