\name{first_order_elimination}
\alias{first_order_elimination}
\title{first_order_elimination}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.first_order_elimination}{Link to Python API reference} (e.g. for correct rendering of equations)


}
\usage{
first_order_elimination(model)
}
\arguments{

}
\value{

}
