% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{fit}
\alias{fit}
\title{fit}
\usage{
fit(
  model_or_models,
  esttool = NULL,
  name = NULL,
  context = NULL,
  ncores = 1,
  validate_dataset = FALSE
)
}
\arguments{
\item{model_or_models}{(Model or array(Model)) List of models or one single model}

\item{esttool}{(str (optional)) Estimation tool to use. NULL to use default}

\item{name}{(str (optional)) Name of run}

\item{context}{(Context (optional)) Run in this context}

\item{ncores}{(numeric) Number of cores to use for estimation}

\item{validate_dataset}{(logical) Whether to validate Pharmpy dataset against FDATA file (relevant for NONMEM models)}
}
\value{
(ModelfitResults | vector of ModelfitResults) ModelfitResults for the model or models
}
\description{
Fit models.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
results <- fit(model)
}
}
\seealso{
run_tool
}
