% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{fix_or_unfix_parameters}
\alias{fix_or_unfix_parameters}
\title{fix_or_unfix_parameters}
\usage{
fix_or_unfix_parameters(model, parameters)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameters}{(list) Set fix/unfix for these parameters}
}
\value{
(Model) Reference to the same model object
}
\description{
Fix or unfix parameters

Set fixedness of parameters to specified values
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$parameters['THETA(1)']
fix_or_unfix_parameters(model, list('THETA(1)'=TRUE))
model$parameters['THETA(1)']
}
}
\seealso{
fix_parameters : Fix parameters

unfix_paramaters : Unfixing parameters

fix_paramaters_to : Fixing parameters and setting a new initial estimate in the same

function

unfix_paramaters_to : Unfixing parameters and setting a new initial estimate in the same

function
}
