% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{generate_model_code}
\alias{generate_model_code}
\title{generate_model_code}
\usage{
generate_model_code(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(str) Model code
}
\description{
Get the model code of the underlying model language
}
\examples{
\dontrun{
model <- load_example_model("pheno")
generate_model_code(model)
}

}
