% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_admid}
\alias{get_admid}
\title{get_admid}
\usage{
get_admid(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(data.frame) ADMID
}
\description{
Get the admid from model dataset

If an administration column is present this will be extracted otherwise
an admid column will be created.
1 : Oral dose
2 : IV dose
}
