% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_baselines}
\alias{get_baselines}
\title{get_baselines}
\usage{
get_baselines(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(data.frame) Dataset with the baselines
}
\description{
Baselines for each subject.

Baseline is taken to be the first row even if that has a missing value.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_baselines(model)
}

}
