% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_bioavailability}
\alias{get_bioavailability}
\title{get_bioavailability}
\usage{
get_bioavailability(model)
}
\arguments{
\item{model}{(Pharmpy model)
Result
list
Dictionary from compartment name to bioavailability expression}
}
\description{
Get bioavailability of doses for all compartments
}
