% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_cmt}
\alias{get_cmt}
\title{get_cmt}
\usage{
get_cmt(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(data.frame) CMT
}
\description{
Get the cmt (compartment) column from the model dataset

If a cmt column is present this will be extracted otherwise
a cmt column will be created.
}
