% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_column_name}
\alias{get_column_name}
\title{get_column_name}
\usage{
get_column_name(model, type)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{type}{(str) Column type. See :py:attr:pharmpy.model.datainfo.ColumnInfo.type}
}
\value{
(str or NULL) Name of the column. NULL if no column found
}
\description{
Retrieve the column with a certain type

If multiple columns have the same type an exception will be raised.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_column_name(model, "dose")
}

}
