% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_concentration_parameters_from_data}
\alias{get_concentration_parameters_from_data}
\title{get_concentration_parameters_from_data}
\usage{
get_concentration_parameters_from_data(model)
}
\arguments{
\item{model}{(Model) Pharmpy model object}
}
\value{
(data.frame) Concentration parameters
}
\description{
Create a dataframe with concentration parameters

Note that all values are directly calculated from the dataset
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_concentration_parameters_from_data(model)
}

}
