% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_config_path}
\alias{get_config_path}
\title{get_config_path}
\usage{
get_config_path()
}
\value{
(str) Path to user config
}
\description{
Returns path to the user config path
}
\examples{
\dontrun{
get_config_path().replace('\\\\', '/')
}

}
