% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_doseid}
\alias{get_doseid}
\title{get_doseid}
\usage{
get_doseid(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(data.frame) DOSEIDs
}
\description{
Get a DOSEID series from the dataset with an id of each dose period starting from 1

If a a dose and observation exist at the same time point the observation will be counted
towards the previous dose.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_doseid(model)
}

}
