% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_dv_symbol}
\alias{get_dv_symbol}
\title{get_dv_symbol}
\usage{
get_dv_symbol(model, dv = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{dv}{(str or numeric (optional)) Either a dv symbol, str or dvid. If NULL (default) return the
only or first dv.}
}
\value{
(sympy.Symbol) DV symbol
}
\description{
Get the symbol for a certain dvid or dv and check that it is valid
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_dv_symbol(model, "Y")
get_dv_symbol(model, 1)
}

}
