% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_evid}
\alias{get_evid}
\title{get_evid}
\usage{
get_evid(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(data.frame) EVID
}
\description{
Get the evid from model dataset

If an event column is present this will be extracted otherwise
an evid column will be created.
}
