% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_ids}
\alias{get_ids}
\title{get_ids}
\usage{
get_ids(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(vector) All subject ids
}
\description{
Retrieve a vector of all subject ids of the dataset
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_ids(model)
}

}
