% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_individual_prediction_expression}
\alias{get_individual_prediction_expression}
\title{get_individual_prediction_expression}
\usage{
get_individual_prediction_expression(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Expr) Symbolic expression
}
\description{
Get the full symbolic expression for the modelled individual prediction

This function currently only support models without ODE systems
}
\examples{
\dontrun{
model <- load_example_model("pheno_linear")
get_individual_prediction_expression(model)
}
}
\seealso{
get_population_prediction_expression : Get full symbolic epression for the population prediction
}
