% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_model_code}
\alias{get_model_code}
\title{get_model_code}
\usage{
get_model_code(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(str) Model code
}
\description{
Get the model code of the underlying model language as a string
}
\examples{
\dontrun{
model <- load_example_model("pheno")
code <- get_model_code(model)
}

}
