% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_mu_connected_to_parameter}
\alias{get_mu_connected_to_parameter}
\title{get_mu_connected_to_parameter}
\usage{
get_mu_connected_to_parameter(model, parameter)
}
\arguments{
\item{model}{(Model) Pharmpy model object.}

\item{parameter}{(str) Name of parameter which to find Mu parameter for.}
}
\value{
(str) Name of Mu parameter or NULL
}
\description{
Return Mu name connected to parameter

If the given parameter is not dependent on any Mu, NULL is returned
}
