% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_number_of_individuals}
\alias{get_number_of_individuals}
\title{get_number_of_individuals}
\usage{
get_number_of_individuals(model)
}
\arguments{
\item{model}{(Model) Pharmpy model

int
Number of individuals in the model dataset}
}
\description{
Retrieve the number of individuals in the model dataset
}
\note{
For NONMEM models this is the number of individuals of the active dataset, i.e. after filteringof IGNORE and ACCEPT and removal of individuals with no observations.
}
\seealso{
get_number_of_observations : Get the number of observations in a dataset

get_number_of_observations_per_individual : Get the number of observations per individual in a

dataset
}
