% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_number_of_observations_per_individual}
\alias{get_number_of_observations_per_individual}
\title{get_number_of_observations_per_individual}
\usage{
get_number_of_observations_per_individual(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Number of observations for each individual
}
