% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_number_of_transit_compartments}
\alias{get_number_of_transit_compartments}
\title{get_number_of_transit_compartments}
\usage{
get_number_of_transit_compartments(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(integer) Number of transit compartments
}
\description{
Return the number of transit compartments in the model
}
