% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_observation_expression}
\alias{get_observation_expression}
\title{get_observation_expression}
\usage{
get_observation_expression(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Expr) Symbolic expression
}
\description{
Get the full symbolic expression for the observation according to the model

This function currently only support models without ODE systems
}
\examples{
\dontrun{
model <- load_example_model("pheno_linear")
expr <- get_observation_expression(model)
print(expr$unicode())
}

}
