% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_observations}
\alias{get_observations}
\title{get_observations}
\usage{
get_observations(model, keep_index = FALSE, dv = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{keep_index}{(logical) Set to TRUE if the original index should be kept.
Otherwise a new index using ID and idv will be created.}

\item{dv}{(Expr or str or numeric (optional)) Name or DVID of dependent variable. NULL for the default (first or only)}
}
\value{
(data.frame) Observations indexed over ID and TIME
}
\description{
Get observations from dataset
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_observations(model)
}
}
\seealso{
get_number_of_observations : get the number of observations

get_number_of_observations_per_individual : get the number of observations per individual
}
