% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_pd_parameters}
\alias{get_pd_parameters}
\title{get_pd_parameters}
\usage{
get_pd_parameters(model)
}
\arguments{
\item{model}{(Model) Pharmpy model to retrieve the PD parameters from}
}
\value{
(vectorc(str)) A vector of the PD parameter names of the given model
}
\description{
Retrieves PD parameters in :class:\code{pharmpy.model.Model}.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_direct_effect(model, "linear")
get_pd_parameters(model)
}
}
\seealso{
get_pk_parameters
}
