% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_rv_parameters}
\alias{get_rv_parameters}
\title{get_rv_parameters}
\usage{
get_rv_parameters(model, rv)
}
\arguments{
\item{model}{(Model) Pharmpy model to retrieve parameters from}

\item{rv}{(str) Name of random variable to retrieve}
}
\value{
(vectorc(str)) A vector of parameter names for the given random variable
}
\description{
Retrieves parameters in :class:\code{pharmpy.model.Model} given a random variable.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_rv_parameters(model, 'ETA_CL')
}
}
\seealso{
has_random_effect

get_pk_parameters

get_individual_parameters
}
