% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_thetas}
\alias{get_thetas}
\title{get_thetas}
\usage{
get_thetas(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Parameters) All theta parameters
}
\description{
Get all thetas (structural parameters) of a model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_thetas(model)
}
}
\seealso{
get_omegas : Get omega parameters

get_sigmas : Get sigma parameters
}
