% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_zero_order_inputs}
\alias{get_zero_order_inputs}
\title{get_zero_order_inputs}
\usage{
get_zero_order_inputs(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Matrix) Vector of inputs
}
\description{
Get zero order inputs for all compartments
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_zero_order_inputs(model)
}

}
