\name{has_additive_error}
\alias{has_additive_error}
\title{has_additive_error}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.has_additive_error}{Link to Python API reference} (e.g. for correct rendering of equations)

Check if a model has an additive error model
}
\usage{
has_additive_error(model)
}
\arguments{

}
\value{

}
