% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_additive_error_model}
\alias{has_additive_error_model}
\title{has_additive_error_model}
\usage{
has_additive_error_model(model)
}
\arguments{
\item{model}{(Model) The model to check}
}
\description{
Check if a model has an additive error model
}
