\name{has_combined_error}
\alias{has_combined_error}
\title{has_combined_error}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.has_combined_error}{Link to Python API reference} (e.g. for correct rendering of equations)

Check if a model has a combined additive and proportinal error model
}
\usage{
has_combined_error(model)
}
\arguments{

}
\value{

}
