% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_combined_error_model}
\alias{has_combined_error_model}
\title{has_combined_error_model}
\usage{
has_combined_error_model(model)
}
\arguments{
\item{model}{(Model) The model to check}
}
\description{
Check if a model has a combined additive and proportinal error model
}
