% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_first_order_absorption}
\alias{has_first_order_absorption}
\title{has_first_order_absorption}
\usage{
has_first_order_absorption(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Bool : TRUE if model has first order absorption)
}
\description{
Check if ode system describes a first order absorption

Currently defined as the central compartment having a unidirectional input
flow from another compartment (such as depot or transit)
}
