% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_first_order_elimination}
\alias{has_first_order_elimination}
\title{has_first_order_elimination}
\usage{
has_first_order_elimination(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(logical) TRUE if model has describes first order elimination
}
\description{
Check if the model describes first order elimination

This function relies on heuristics and will not be able to detect all
possible ways of coding the first order elimination.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
has_first_order_elimination(model)
}

}
