% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_linear_odes}
\alias{has_linear_odes}
\title{has_linear_odes}
\usage{
has_linear_odes(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(logical) TRUE if model has an ODE system that is linear
}
\description{
Check if model has a linear ODE system
}
\examples{
\dontrun{
model <- load_example_model("pheno")
has_linear_odes(model)
}
}
\seealso{
has_odes

has_linear_odes_with_real_eigenvalues
}
