% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_linear_odes_with_real_eigenvalues}
\alias{has_linear_odes_with_real_eigenvalues}
\title{has_linear_odes_with_real_eigenvalues}
\usage{
has_linear_odes_with_real_eigenvalues(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(logical) TRUE if model has an ODE system that is linear
}
\description{
Check if model has a linear ode system with real eigenvalues
}
\examples{
\dontrun{
model <- load_example_model("pheno")
has_linear_odes_with_real_eigenvalues(model)
}
}
\seealso{
has_odes

has_linear_odes
}
