% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_mu_reference}
\alias{has_mu_reference}
\title{has_mu_reference}
\usage{
has_mu_reference(model)
}
\arguments{
\item{model}{(Model) Pharmpy model object}
}
\value{
(logical) Whether the model is mu referenced
}
\description{
Check if model is Mu-reference or not.

Will return TRUE if each parameter with an ETA is dependent on a Mu parameter.
}
