\name{has_proportional_error}
\alias{has_proportional_error}
\title{has_proportional_error}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.has_proportional_error}{Link to Python API reference} (e.g. for correct rendering of equations)

Check if a model has a proportional error model
}
\usage{
has_proportional_error(model)
}
\arguments{

}
\value{

}
