% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_seq_zo_fo_absorption}
\alias{has_seq_zo_fo_absorption}
\title{has_seq_zo_fo_absorption}
\usage{
has_seq_zo_fo_absorption(model)
}
\arguments{
\item{model}{(Model) DPharmpy model}
}
\description{
Check if ode system describes a sequential zero-order, first-order absorption

Defined as the model having both zero- and first-order absorption.
}
\seealso{
has_zero_order_absorption

has_first_order_absorption
}
