% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_weibull_absorption}
\alias{has_weibull_absorption}
\title{has_weibull_absorption}
\usage{
has_weibull_absorption(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Bool : TRUE if model has weibull type absorption)
}
\description{
Check if ode system describes a weibull type absorption
}
