% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_zero_order_absorption}
\alias{has_zero_order_absorption}
\title{has_zero_order_absorption}
\usage{
has_zero_order_absorption(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(logical) Whether the model has zero order absorption or not
}
\description{
Check if ode system describes a zero order absorption

currently defined as having Infusion dose with rate not in dataset
}
\examples{
\dontrun{
model <- load_example_model("pheno")
has_zero_order_absorption(model)
}

}
