% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_zero_order_elimination}
\alias{has_zero_order_elimination}
\title{has_zero_order_elimination}
\usage{
has_zero_order_elimination(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(logical) TRUE if model has describes zero order elimination
}
\description{
Check if the model describes zero-order elimination

This function relies on heuristics and will not be able to detect all
possible ways of coding the zero-order elimination.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
has_zero_order_elimination(model)
set_zero_order_elimination(model)
has_zero_order_elimination(model)
}

}
