% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{infer_datatypes}
\alias{infer_datatypes}
\title{infer_datatypes}
\usage{
infer_datatypes(model, columns = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{columns}{(array(str) (optional)) The columns to attempt conversion or NULL for all}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Infer and set datatypes for the dataset

All columns or only the ones set in the columns option will be checked
for conversion to a simpler datatype. Currently only int32 will be checked since
this corresponds to the integer type in R.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- infer_datatypes(model, c('APGR'))
model$dataset
}

}
