% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{insert_ebes_into_dataset}
\alias{insert_ebes_into_dataset}
\title{insert_ebes_into_dataset}
\usage{
insert_ebes_into_dataset(
  model,
  individual_estimates,
  individual_estimates_covariance = NULL
)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{individual_estimates}{(data.frame) Individual eta estimates (EBEs). Could be taken directly from ModelfitResults.}

\item{individual_estimates_covariance}{(data.frame (optional)) Uncertainties of individual estimates (ETCs). Could be taken directly from ModelfitResults.}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Insert EBEs and ETCs from results into the dataset of a model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
results <- load_example_modelfit_results("pheno")
ebes <- results$individual_estimates
etcs <- results$individual_estimates_covariance
model2 <- insert_ebes_into_dataset(model, ebes, etcs)
model2$datainfo$names
}

}
