% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{is_binary}
\alias{is_binary}
\title{is_binary}
\usage{
is_binary(model, expr)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{expr}{(str) A mathematical expression containing column or variable names}
}
\value{
(logical) TRUE if binary
}
\description{
Check if an expression defines a binary variable (only having values 0 and 1)
}
\examples{
\dontrun{
model <- load_example_model("pheno")
is_binary(model, "WGT")
is_binary(model, "FA1")
}

}
