% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{is_linearized}
\alias{is_linearized}
\title{is_linearized}
\usage{
is_linearized(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(logical) TRUE if model has been linearized and FALSE otherwise
}
\description{
Determine if a model is linearized
}
\examples{
\dontrun{
model1 <- load_example_model("pheno")
is_linearized(model1)
model2 <- load_example_model("pheno_linear")
is_linearized(model2)
}

}
