% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{is_real}
\alias{is_real}
\title{is_real}
\usage{
is_real(model, expr)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{expr}{(numeric or str or Expr) Expression to test}
}
\value{
(logical or NULL) TRUE if expression is real, FALSE if not and NULL if unknown
}
\description{
Determine if an expression is real valued given constraints of a model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
is_real(model, "CL")
}

}
