% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{is_simulation_model}
\alias{is_simulation_model}
\title{is_simulation_model}
\usage{
is_simulation_model(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(logical) TRUE if it is a simulation model
}
\description{
Check if a model is a pure simulation model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
is_simulation_model(model)
}

}
