% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{load_dataset}
\alias{load_dataset}
\title{load_dataset}
\usage{
load_dataset(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Load the dataset given datainfo
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- unload_dataset(model)
model$dataset is NULL
model <- load_dataset(model)
model$dataset
}

}
