% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{load_example_model}
\alias{load_example_model}
\title{load_example_model}
\usage{
load_example_model(name)
}
\arguments{
\item{name}{(str) Name of the model. Currently available models are "pheno" and "pheno_linear"}
}
\value{
(Model) Pharmpy model
}
\description{
Load an example model

Load an example model from models built into Pharmpy
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$statements
}

}
