% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{load_example_modelfit_results}
\alias{load_example_modelfit_results}
\title{load_example_modelfit_results}
\usage{
load_example_modelfit_results(name)
}
\arguments{
\item{name}{(str) Name of the model. Currently available models are "pheno" and "pheno_linear"}
}
\value{
(ModelfitResults) Loaded modelfit results object
}
\description{
Load the modelfit results of an example model

Load the modelfit results of an example model built into Pharmpy
}
\examples{
\dontrun{
results <- load_example_modelfit_results("pheno")
results$parameter_estimates
}

}
