% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{make_declarative}
\alias{make_declarative}
\title{make_declarative}
\usage{
make_declarative(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Pharmpy model object
}
\description{
Make the model statments declarative

Each symbol will only be declared once.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$statements$before_odes
model <- make_declarative(model)
model$statements$before_odes
}

}
