% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{map_eta_parameters}
\alias{map_eta_parameters}
\title{map_eta_parameters}
\usage{
map_eta_parameters(model, keys, values, level = "iiv")
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{keys}{(str) What to map from. Either parameters, omegas or etas}

\item{values}{(str) What to map to. Either parameters, omegas or etas}

\item{level}{(str) Which variability level to consider. Either iiv or iov. iiv is the default.}
}
\value{
(list) A list from parameters, omegas or etas to a vector of connected parameters, omegas or etas.
}
\description{
Create a map with the connections from and to individual parameters, omegas and/or etas

The mapping will always be one to many.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
map_eta_parameters(model, "parameters", "omegas")
}

}
