\name{michaelis_menten_elimination}
\alias{michaelis_menten_elimination}
\title{michaelis_menten_elimination}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.michaelis_menten_elimination}{Link to Python API reference} (e.g. for correct rendering of equations)

Sets elimination to Michaelis-Menten. Initial estimate for CLMM is set to CL and KM is set to

:math:`2*max(DV)`.
}
\usage{
michaelis_menten_elimination(model)
}
\arguments{

}
\value{

}
