\name{mixed_mm_fo_elimination}
\alias{mixed_mm_fo_elimination}
\title{mixed_mm_fo_elimination}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.mixed_mm_fo_elimination}{Link to Python API reference} (e.g. for correct rendering of equations)

Sets elimination to mixed Michaelis-Menten and first order. Initial estimate for CLMM is set

to CL/2 and KM is set to :math:`2*max(DV)`.
}
\usage{
mixed_mm_fo_elimination(model)
}
\arguments{

}
\value{

}
