% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{mu_reference_model}
\alias{mu_reference_model}
\title{mu_reference_model}
\usage{
mu_reference_model(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Convert model to use mu-referencing

Mu-referencing an eta is to separately define its actual mu (mean) parameter.
For example: (equation could not be rendered, see API doc on website)
normal distribution would give (equation could not be rendered, see API doc on website)
(equation could not be rendered, see API doc on website)
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- mu_reference_model(model)
model$statements$before_odes
}

}
