\name{ninds}
\alias{ninds}
\title{ninds}
\description{
Retrieve the number of individuals in the model dataset.

\bold{Usage}

\code{ninds(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.ninds}{API reference} (for correct rendering of equations, tables etc.).
Retrieve the number of individuals in the model dataset

This documentation was automatically generated from Pharmpy (0.26.0).}
