\name{nobs}
\alias{nobs}
\title{nobs}
\description{
Retrieve the total number of observations in the model dataset.

\bold{Usage}

\code{nobs(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.nobs}{API reference} (for correct rendering of equations, tables etc.).
Retrieve the total number of observations in the model dataset

This documentation was automatically generated from Pharmpy (0.25.0).}
