\name{nobsi}
\alias{nobsi}
\title{nobsi}
\description{
Number of observations for each individual.

\bold{Usage}

\code{nobsi(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.nobsi}{API reference} (for correct rendering of equations, tables etc.).
Number of observations for each individual

This documentation was automatically generated from Pharmpy (0.25.1).}
