% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{omit_data}
\alias{omit_data}
\title{omit_data}
\usage{
omit_data(dataset_or_model, group, name_pattern = "omitted_{}")
}
\arguments{
\item{dataset_or_model}{(data.frame or Model) Dataset or model for which to omit records}

\item{group}{(str) Name of the column to use for grouping}

\item{name_pattern}{(str) Name to use for generated datasets. A number starting from 1 will be put in the placeholder.}
}
\value{
(iterator) Iterator yielding tuples of models/dataframes and the omitted group
}
\description{
Iterate over omissions of a certain group in a dataset. One group is omitted at a time.
}
