% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{open_context}
\alias{open_context}
\title{open_context}
\usage{
open_context(name, ref = NULL)
}
\arguments{
\item{name}{(str) Name of the context}

\item{ref}{(str (optional)) Parent path of the context}
}
\description{
Open a context from a tool run
}
\examples{
\dontrun{
ctx <- open_context("myrun")
}

}
