% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{plot_eta_distributions}
\alias{plot_eta_distributions}
\title{plot_eta_distributions}
\usage{
plot_eta_distributions(model, individual_estimates)
}
\arguments{
\item{model}{(Model) Previously run Pharmpy model.}

\item{individual_estimates}{(data.frame) Individual estimates for etas}
}
\value{
(alt.Chart) Plot
}
\description{
Plot eta distributions for all etas
}
\examples{
\dontrun{
model <- load_example_model("pheno")
res <- load_example_modelfit_results("pheno")
plot_eta_distributions(model, res$individual_estimates)
}

}
