% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{plot_individual_predictions}
\alias{plot_individual_predictions}
\title{plot_individual_predictions}
\usage{
plot_individual_predictions(model, predictions = NULL, individuals = NULL)
}
\arguments{
\item{model}{(Model) Previously run Pharmpy model.}

\item{predictions}{(vector) A vector of names of predictions to plot. NULL for all available}

\item{individuals}{(vector) A vector of individuals to include. NULL for all individuals}
}
\value{
(alt.Chart) Plot
}
\description{
Plot DV and predictions grouped on individuals
}
