% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{plot_iofv_vs_iofv}
\alias{plot_iofv_vs_iofv}
\title{plot_iofv_vs_iofv}
\usage{
plot_iofv_vs_iofv(iofv1, iofv2, name1, name2)
}
\arguments{
\item{iofv1}{(array) Estimated iOFV of the first model}

\item{iofv2}{(array) Estimated iOFV of the second model}

\item{name1}{(str) Name of first model}

\item{name2}{(str) Name of second model}
}
\value{
(alt.Chart) Scatterplot
}
\description{
Plot individual OFV of two models against each other
}
\examples{
\dontrun{
}

}
