% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{plot_transformed_eta_distributions}
\alias{plot_transformed_eta_distributions}
\title{plot_transformed_eta_distributions}
\usage{
plot_transformed_eta_distributions(
  model,
  parameter_estimates,
  individual_estimates,
  parameter_estimates_untransformed,
  seed = 1234
)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameter_estimates}{(array or list(str=numeric)) Parameter estimates of model fit}

\item{individual_estimates}{(data.frame) Individual estimates for etas}

\item{parameter_estimates_untransformed}{(array or list(str=numeric)) Parameter estimates of untransformed model fit}

\item{seed}{(numeric) Random number generator or seed}
}
\value{
(alt.Chart) Plot
}
\description{
Plot transformed eta distributions for all transformed etas
}
