% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{print_model_code}
\alias{print_model_code}
\title{print_model_code}
\usage{
print_model_code(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Print the model code of the underlying model language to the console
}
\examples{
\dontrun{
model <- load_example_model("pheno")
print_model_code(model)
}

}
