% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{print_model_symbols}
\alias{print_model_symbols}
\title{print_model_symbols}
\usage{
print_model_symbols(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Print all symbols defined in a model

Symbols will be in one of the categories thetas, etas, omegas, epsilons, sigmas,
variables and data columns
}
\examples{
\dontrun{
model <- load_example_model("pheno")
print_model_symbols(model)
}

}
