\name{proportional_error}
\alias{proportional_error}
\title{proportional_error}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.proportional_error}{Link to Python API reference} (e.g. for correct rendering of equations)

Set a proportional error model. Initial estimate for new sigma is 0.09.
}
\usage{
proportional_error(model)
}
\arguments{

}
\value{

}
