% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{read_datainfo}
\alias{read_datainfo}
\title{read_datainfo}
\usage{
read_datainfo(path)
}
\arguments{
\item{path}{(str) A path to a datainfo file}
}
\value{
(DataInfo) DataInfo object
}
\description{
Read a datainfo file
}
