% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{read_dataset_from_datainfo}
\alias{read_dataset_from_datainfo}
\title{read_dataset_from_datainfo}
\usage{
read_dataset_from_datainfo(datainfo)
}
\arguments{
\item{datainfo}{(DataInfo | Path | str) A datainfo object or a path to a datainfo object}
}
\value{
(data.frame) The dataset
}
\description{
Read a dataset given a datainfo object or path to a datainfo file
}
