\name{read_model}
\alias{read_model}
\title{read_model}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.read_model}{Link to Python API reference} (e.g. for correct rendering of equations)

Read model from file
}
\usage{
read_model(path)
}
\arguments{

}
\value{

}
