% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{read_model_from_database}
\alias{read_model_from_database}
\title{read_model_from_database}
\usage{
read_model_from_database(name, database = NULL)
}
\arguments{
\item{name}{(str) Name of model to use as lookup}

\item{database}{(Database) Database to use. Will use default database if not specified.}
}
\value{
(Model) Read model object
}
\description{
Read model from model database
}
\examples{
\dontrun{
model <- read_model_from_database("run1")
}
}
\seealso{
read_model : Read model from file

read_model_from_string : Read model from string
}
