\name{read_model_from_string}
\alias{read_model_from_string}
\title{read_model_from_string}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.read_model_from_string}{Link to Python API reference} (e.g. for correct rendering of equations)

Read model directly from the model code in a string
}
\usage{
read_model_from_string(code)
}
\arguments{

}
\value{

}
