% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{read_results}
\alias{read_results}
\title{read_results}
\usage{
read_results(path)
}
\arguments{
\item{path}{(str, Path) Path to results file}
}
\value{
(Results) Results object for tool
}
\description{
Read results object from file
}
\examples{
\dontrun{
res <- read_results("results$json")
}
}
\seealso{
create_results
}
