% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_bioavailability}
\alias{remove_bioavailability}
\title{remove_bioavailability}
\usage{
remove_bioavailability(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Remove bioavailability from the first dose compartment of model.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- remove_bioavailability(model)
}
}
\seealso{
set_bioavailability
}
