% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_covariance_step}
\alias{remove_covariance_step}
\title{remove_covariance_step}
\usage{
remove_covariance_step(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Reference to the same model object
}
\description{
Removes covariance step to the final estimation step
}
\examples{
\dontrun{
model <- load_example_model("pheno")
remove_covariance_step(model)
ests <- model$estimation_steps
ests[1]
}
}
\seealso{
add_estimation_step

set_estimation_step

remove_estimation_step

append_estimation_step_options

add_covariance_step

set_evaluation_step
}
