\name{remove_error}
\alias{remove_error}
\title{remove_error}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.remove_error}{Link to Python API reference} (e.g. for correct rendering of equations)

Remove error model.
}
\usage{
remove_error(model)
}
\arguments{

}
\value{

}
