\name{remove_estimation_step}
\alias{remove_estimation_step}
\title{remove_estimation_step}
\description{
Remove estimation step.

\bold{Usage}

\code{remove_estimation_step(model, est_idx)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.remove_estimation_step}{API reference} (for correct rendering of equations, tables etc.).
Remove estimation step

This documentation was automatically generated from Pharmpy (0.26.1).}
\arguments{

\item{
model
}{
Model. Pharmpy model
}
\item{
est_idx
}{
numeric. index of estimation step to remove
}
}
\value{
model : Model
}
