\name{remove_iiv}
\alias{remove_iiv}
\title{remove_iiv}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.remove_iiv}{Link to Python API reference} (e.g. for correct rendering of equations)

Removes all IIV omegas given a list with eta names and/or parameter names.
}
\usage{
remove_iiv(model, list_to_remove=None)
}
\arguments{

\item{
model
}{
Model, Pharmpy model to create block effect on.
}
\item{
list_to_remove
}{
list, List of etas and/or parameter name to remove. If None, all etas that are IIVs will
}
}
\value{

}
