\name{remove_lag_time}
\alias{remove_lag_time}
\title{remove_lag_time}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.remove_lag_time}{Link to Python API reference} (e.g. for correct rendering of equations)

Remove lag time from the dose compartment of model.
}
\usage{
remove_lag_time(model)
}
\arguments{

}
\value{

}
