% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_loq_data}
\alias{remove_loq_data}
\title{remove_loq_data}
\usage{
remove_loq_data(model, lloq = NULL, uloq = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model object}

\item{lloq}{(numeric (optional)) Lower limit of quantification. Default not specified.}

\item{uloq}{(numeric (optional)) Upper limit of quantification. Default not specified.}
}
\value{
(Model) Reference to the same model object
}
\description{
Remove loq data records from the dataset

Does nothing if none of the limits is specified.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
remove_loq_data(model, lloq=10, uloq=40)
length(model$dataset)
}

}
