% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_parameter_uncertainty_step}
\alias{remove_parameter_uncertainty_step}
\title{remove_parameter_uncertainty_step}
\usage{
remove_parameter_uncertainty_step(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Pharmpy model object
}
\description{
Removes parameter uncertainty step from the final estimation step
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- remove_parameter_uncertainty_step(model)
ests <- model$estimation_steps
ests[1]
}
}
\seealso{
add_estimation_step

set_estimation_step

remove_estimation_step

append_estimation_step_options

add_parameter_uncertainty_step

set_evaluation_step
}
