% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_predictions}
\alias{remove_predictions}
\title{remove_predictions}
\usage{
remove_predictions(model, to_remove = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{to_remove}{(array(str) (optional)) Predictions to remove}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Remove predictions and/or residuals

Remove predictions from estimation step. Default is to remove all predictions.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- remove_predictions(model)
model$execution_steps[-1].predictions
}
}
\seealso{
add_predictions

add_residuals

set_estimation_step

add_estimation_step

remove_estimation_step

append_estimation_step_options

add_parameter_uncertainty_step

remove_parameter_uncertainty_step
}
