% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_residuals}
\alias{remove_residuals}
\title{remove_residuals}
\usage{
remove_residuals(model, to_remove = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{to_remove}{(array(str)) List of predictions to remove}
}
\value{
(Model) Pharmpy model object
}
\description{
Remove predictions and/or residuals

Remove residuals from estimation step.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- remove_residuals(model, 'all')
model$estimation_steps[-1].residuals
}
}
\seealso{
add_predictions

add_residuals

set_estimation_step

add_estimation_step

remove_estimation_step

append_estimation_step_options

add_parameter_uncertainty_step

remove_parameter_uncertainty_step
}
