% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_unused_columns}
\alias{remove_unused_columns}
\title{remove_unused_columns}
\usage{
remove_unused_columns(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Remove all columns in the dataset that are not used by the model including dropped columns

Warnings
Currently columns not needed to give a prediction, but used by other expressions in the
model are kept. This will change in the future.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- remove_unused_columns(model)
vector(model$dataset$columns)
}
}
\seealso{
drop_columns : Drop columns from the dataset
}
