% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{replace_fixed_thetas}
\alias{replace_fixed_thetas}
\title{replace_fixed_thetas}
\usage{
replace_fixed_thetas(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Replace all fixed thetas with constants in the model statements
}
