% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{reset_dataset}
\alias{reset_dataset}
\title{reset_dataset}
\usage{
reset_dataset(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Use Pharmpy dataset instead of original file

The parsed dataset will be written down when doing e.g. estimation or writing files
instead of the original dataset file

Removes path and provenance from DataInfo.

Note: Use this to get same behavior as Pharmpy<=2.1.1
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- reset_dataset(model)
}

}
