% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reset_index}
\alias{reset_index}
\title{Reset index}
\usage{
reset_index(df)
}
\arguments{
\item{df}{A data.frame coming from Pharmpy}
}
\description{
Reset index of dataframe.

Reset index from a multi indexed data.frame or a non-standard index
}
