% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{reset_indices_results}
\alias{reset_indices_results}
\title{Reset result indices}
\usage{
reset_indices_results(res)
}
\arguments{
\item{res}{A Pharmpy results object}
}
\description{
Resets indices in dataframes within Results-objects when needed
}
