% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{resume_tool}
\alias{resume_tool}
\title{resume_tool}
\usage{
resume_tool(path)
}
\arguments{
\item{path}{(str) The path to the tool database}
}
\value{
(Results) Results object for tool
}
\description{
Resume tool workflow from tool database path
}
\examples{
\dontrun{
res <- resume_tool("resmod_dir1")
}

}
