% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{retrieve_final_model}
\alias{retrieve_final_model}
\title{retrieve_final_model}
\usage{
retrieve_final_model(res)
}
\arguments{
\item{res}{(Results) A results object}
}
\value{
(Model) Reference to final model
}
\description{
Retrieve final model from a result object
}
\examples{
\dontrun{
res <- read_results("results$json")
model <- retrieve_final_model(res)
}
}
\seealso{
retrieve_models
}
