% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{retrieve_model}
\alias{retrieve_model}
\title{retrieve_model}
\usage{
retrieve_model(context, name)
}
\arguments{
\item{context}{(Context) A previously opened context}

\item{name}{(str) Name of the model or a qualified name with a subcontext path, e.g.}
}
\value{
(Model) The model object
}
\description{
Retrieve a model from a context

Any models created and run by the tool can be
retrieved.
}
\examples{
\dontrun{
context <- open_context(ref='path/to/', name='modelsearch1')
model <- retrieve_model(context, 'run1')
}

}
