% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{retrieve_modelfit_results}
\alias{retrieve_modelfit_results}
\title{retrieve_modelfit_results}
\usage{
retrieve_modelfit_results(source, name)
}
\arguments{
\item{source}{(str or Context) Source where to find models. Can be a path (as str or Path), or a
Context}

\item{name}{(str) Name of the model}
}
\value{
(ModelfitResults) The results object
}
\description{
Retrieve the modelfit results of a model
}
\examples{
\dontrun{
tooldir_path <- 'path/to/tool/directory'
context <- create_context("iivsearch1")
results <- retrieve_modelfit_results(context, 'input')
}

}
