% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{retrieve_models}
\alias{retrieve_models}
\title{retrieve_models}
\usage{
retrieve_models(path, names = NULL)
}
\arguments{
\item{path}{(str or Path) A path to the tool directory}

\item{names}{(vector) List of names of the models to retrieve or NULL for all}
}
\value{
(vector) List of retrieved model objects
}
\description{
Retrieve models after a tool runs

Any models created and run by the tool can be
retrieved.
}
