% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_bootstrap}
\alias{run_bootstrap}
\title{run_bootstrap}
\usage{
run_bootstrap(model, results = NULL, resamples = 1, dofv = FALSE, ...)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{results}{(ModelfitResults (optional)) Results for model}

\item{resamples}{(numeric) Number of bootstrap resamples}

\item{dofv}{(logical) Will evaluate bootstrap models with original dataset if se}

\item{...}{Arguments to pass to tool}
}
\value{
(BootstrapResults) Bootstrap tool result object
}
\description{
Run bootstrap tool
}
\examples{
\dontrun{
model <- load_example_model("pheno")
results <- load_example_modelfit_results("pheno")
run_bootstrap(model, res, resamples=500)
}

}
