% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_iiv}
\alias{run_iiv}
\title{run_iiv}
\usage{
run_iiv(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Reference to the same model object
}
\description{
Run IIV tool

Runs two IIV workflows: testing the number of etas and testing which block structure
}
\examples{
\dontrun{
model <- load_example_model("pheno")
run_iiv(model)
}
}
\seealso{
run_amd

run_tool
}
