% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_linearize}
\alias{run_linearize}
\title{run_linearize}
\usage{
run_linearize(model, results, model_name = "linbase", description = "", ...)
}
\arguments{
\item{model}{(Model) Pharmpy model. Should use FOCEI or other EM-methods for estimation}

\item{results}{(ModelfitResults) Results of estimation of model}

\item{model_name}{(str) New name of linearized model. The default is "linbase".}

\item{description}{(str) Description of linearized model. The default is ""}

\item{...}{Arguments to pass to tool}
}
\value{
(LinearizeResults) Linearize tool results object.
}
\description{
Linearize a model
}
