% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_modelrank}
\alias{run_modelrank}
\title{run_modelrank}
\usage{
run_modelrank(
  models,
  results,
  ref_model,
  strictness = "minimization_successful or (rounding_errors and sigdigs >= 0.1)",
  rank_type = "ofv",
  alpha = 0.05,
  search_space = NULL,
  E = NULL,
  parameter_uncertainty_method = NULL,
  ...
)
}
\arguments{
\item{models}{(array(Model)) Models to rank}

\item{results}{(array(ModelfitResults)) Modelfit results to rank on}

\item{ref_model}{(Model) Model to compare to}

\item{strictness}{(str) Strictness criteria}

\item{rank_type}{(str) Which ranking type should be used. Supported types are OFV, LRT, AIC, BIC (mixed, IIV,
random), and mBIC (mixed, IIV, random). Default is OFV.}

\item{alpha}{(numeric (optional)) Cutoff p-value that is considered significant in likelihood ratio test. Default
is NULL}

\item{search_space}{(str or ModelFeatures (optional)) Search space to test. Either as a string or a ModelFeatures object.}

\item{E}{(numeric or str or list(numeric or str,numeric or str) (optional)) Expected number of predictors (used for mBIC). Must be set when using mBIC. Tuple
if mBIC for IIV (both diagonals and off-diagonals)}

\item{parameter_uncertainty_method}{(str (optional)) Parameter uncertainty method. Will be used in ranking models if strictness includes
parameter uncertaint}

\item{...}{Arguments to pass to tool}
}
\value{
(ModelRankResults) ModelRank tool result object
}
\description{
Run ModelRank tool.
}
