% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_resmod}
\alias{run_resmod}
\title{run_resmod}
\usage{
run_resmod(model = NULL, groups = 4, p_value = 0.05, skip = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{groups}{(integer) The number of bins to use for the time varying models}

\item{p_value}{(numeric) The p-value to use for the likelihood ratio test}

\item{skip}{(vector) A vector of models to not attempt.}
}
\value{
(ResmodResults) Resmod tool result object
}
\description{
Run the resmod tool. For more details, see :ref:\code{resmod}.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
run_resmod(model=model)
}

}
