% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_simulation}
\alias{run_simulation}
\title{run_simulation}
\usage{
run_simulation(model = NULL, ...)
}
\arguments{
\item{model}{(Model (optional)) Pharmpy mode}

\item{...}{Arguments to pass to tool}
}
\value{
(SimulationResult) SimulationResults object
}
\description{
Run the simulation tool.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_simulation(model, n=10)
run_simulations(model)
}

}
