% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_tool}
\alias{run_tool}
\title{run_tool}
\usage{
run_tool(tool_name, ...)
}
\arguments{
\item{tool_name}{(str) Name of tool to run}

\item{...}{Arguments to pass to tool}
}
\value{
(Results) Results object for tool
}
\description{
Run tool workflow

note::
This is a general function that can run any tool. There is also one function for each
specific tool. Please refer to the documentation of these for more specific information.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
res <- run_tool("ruvsearch", model)
}

}
