% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_vpc}
\alias{run_vpc}
\title{run_vpc}
\usage{
run_vpc(model, results = NULL, samples = 20, stratify = NULL, ...)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{results}{(ModelfitResults (optional)) Results for model}

\item{samples}{(numeric) Number of samples}

\item{stratify}{(str (optional)) Column to stratify o}

\item{...}{Arguments to pass to tool}
}
\value{
(VPCResults) VPC results object
}
\description{
Run VPC
}
\examples{
\dontrun{
model <- load_example_model("pheno")
res <- load_example_modelfit_results("pheno")
run_vpc(model, res, samples=200)
}

}
