% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{sample_parameters_uniformly}
\alias{sample_parameters_uniformly}
\title{sample_parameters_uniformly}
\usage{
sample_parameters_uniformly(
  model,
  fraction = 0.1,
  parameters = NULL,
  force_posdef_samples = NULL,
  n = 1,
  rng = NULL
)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{fraction}{(float) Fraction of estimate value to use for distribution bounds}

\item{parameters}{(pd.Series) Names of parameters to use. Default is to use all parameters in the model.}

\item{force_posdef_samples}{(int) Number of samples to reject before forcing variability parameters to give
positive definite covariance matrices.}

\item{n}{(int) Number of samples}

\item{rng}{(int or rng) Random number generator or seed

pd.DataFrame : samples}
}
\description{
Sample parameter vectors using uniform sampling

Each parameter value will be randomly sampled from a uniform distribution
with the bounds being estimate ± estimate * fraction.
}
\seealso{
sample_parameters_from_covariance_matrix : Sample parameter vectors using the

uncertainty covariance matrix

sample_individual_estimates : Sample individual estiates given their covariance
}
