\name{seq_zo_fo_absorption}
\alias{seq_zo_fo_absorption}
\title{seq_zo_fo_absorption}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.seq_zo_fo_absorption}{Link to Python API reference} (e.g. for correct rendering of equations)

Set or change to sequential zero order first order absorption rate. Initial estimate for

absorption rate is set the previous rate if available, otherwise it is set to the time of

first observation/2 is used.
}
\usage{
seq_zo_fo_absorption(model)
}
\arguments{

}
\value{

}
