% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_additive_error_model}
\alias{set_additive_error_model}
\title{set_additive_error_model}
\usage{
set_additive_error_model(model, data_trans = NULL)
}
\arguments{
\item{model}{(Model) Set error model for this model}

\item{data_trans}{(str or expression) A data transformation expression or NULL (default) to use the transformation
specified by the model.}
}
\description{
Set an additive error model. Initial estimate for new sigma is :math:\verb{(min(DV)/2)²}.

The error function being applied depends on the data transformation.

+------------------------+----------------------------------------+
| Data transformation    | Additive error                         |
+========================+========================================+
| :math:\code{y}              | :math:\code{f + epsilon_1}                 |
+------------------------+----------------------------------------+
| :math:\code{log(y)}         | :math:\verb{log(f) + frac\{epsilon_1\}\{f\}} |
+------------------------+----------------------------------------+
}
