% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_baseline_effect}
\alias{set_baseline_effect}
\title{set_baseline_effect}
\usage{
set_baseline_effect(model, expr = "const")
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{expr}{(str) Name of baseline effect function.}
}
\value{
(Model) Pharmpy model object
}
\description{
Create baseline effect model.

Currently implemented baseline effects are:

Constant baseline effect (const):

(equation could not be rendered, see API doc on website)
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_baseline_effect(model, expr='const')
model$statements$find_assignment("E")
}

}
