% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_covariates}
\alias{set_covariates}
\title{set_covariates}
\usage{
set_covariates(model, covariates)
}
\arguments{
\item{model}{(Model or DataInfo) Pharmpy model or a DataInfo object}

\item{covariates}{(array(str)) Column names}
}
\value{
(Model) Updated Pharmpy model or DataInfo object
}
\description{
Set columns in the dataset to be covariates in the datainfo
}
