% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_dataset}
\alias{set_dataset}
\title{set_dataset}
\usage{
set_dataset(model, path_or_df, format = NULL, datatype = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{path_or_df}{(str or data.frame) Dataset path or dataframe}

\item{format}{(str (optional)) File format of dataset (default csv. Can also be set to 'nonmem')}

\item{datatype}{(any) Deprecated. Use format instead}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Load the dataset given datainfo
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- unload_dataset(model)
dataset_path <- model$datainfo$path
model$dataset is NULL
model <- set_dataset(model, dataset_path, format='nonmem')
model$dataset
}

}
