% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_description}
\alias{set_description}
\title{set_description}
\usage{
set_description(model, new_description)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{new_description}{(str) New description of model}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Set description of model object
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$description
model <- set_description(model, "PHENOBARB run 2")
model$description
}

}
