% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_direct_effect}
\alias{set_direct_effect}
\title{set_direct_effect}
\usage{
set_direct_effect(model, expr)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{expr}{(str) Name of PD effect function. The function can either be
\itemize{
\item baseline: E = E0
\item step effect: Emax = theta if C > 0 else 0, E = E0 + Emax
\item linear: E = E0 + S * C
\item Emax: E = E0 + Emax * C / (EC50 + C)
\item sigmoidal: E = Emax * C^n / (EC50^n + C^n)
\item log-lin: E = m * log(C + C0)
Valid strings are: baseline, linear, Emax, sigmoid, step, loglin
}}
}
\value{
(Model) Pharmpy model object
}
\description{
Add an effect to a model
}
\examples{
\dontrun{
}

}
