\name{set_dtbs_error}
\alias{set_dtbs_error}
\title{set_dtbs_error}
\description{
Dynamic transform both sides.

\bold{Usage}

\code{set_dtbs_error(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.set_dtbs_error}{API reference} (for correct rendering of equations, tables etc.).
Dynamic transform both sides

This documentation was automatically generated from Pharmpy (0.25.1).}
