% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_dtbs_error_model}
\alias{set_dtbs_error_model}
\title{set_dtbs_error_model}
\usage{
set_dtbs_error_model(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Reference to the same model
}
\description{
Dynamic transform both sides
}
\examples{
\dontrun{
model <- load_example_model("pheno")
set_dtbs_error_model(model)
}

}
