% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_dvid}
\alias{set_dvid}
\title{set_dvid}
\usage{
set_dvid(model, name)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{name}{(str) Name of DVID column}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Set a column to act as DVID. Replace DVID if one is already set.
}
