% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_first_order_elimination}
\alias{set_first_order_elimination}
\title{set_first_order_elimination}
\usage{
set_first_order_elimination(model)
}
\arguments{
\item{model}{(Model) Pharmpy model

Model
Reference to same model}
}
\description{
Sets elimination to first order
}
\seealso{
set_zero_order_elimination

set_michaelis_menten_elimination
}
