% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_index}
\alias{set_index}
\title{Set index}
\usage{
set_index(df, colnames)
}
\arguments{
\item{df}{A data.frame}

\item{colnames}{An array of column names}
}
\description{
Set a multiindex for a dataframe.

The index columns will be used when using the data.frame in a pharmpy call
}
