\name{set_initial_estimates}
\alias{set_initial_estimates}
\title{set_initial_estimates}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.set_initial_estimates}{Link to Python API reference} (e.g. for correct rendering of equations)

Set initial estimates
}
\usage{
set_initial_estimates(model, inits)
}
\arguments{

}
\value{
model: Model
}
