% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_instantaneous_absorption}
\alias{set_instantaneous_absorption}
\title{set_instantaneous_absorption}
\usage{
set_instantaneous_absorption(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Set or change to instantaneous absorption rate.

Currently lagtime together with instantaneous absorption is not supported.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_instantaneous_absorption(model)
model$statements$ode_system
}
}
\seealso{
set_zero_order_absorption

set_first_order_absorption
}
