% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_lag_time}
\alias{set_lag_time}
\title{set_lag_time}
\usage{
set_lag_time(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Add lag time to the dose compartment of model. Initial estimate for lag time is set the
previous lag time if available, otherwise it is set to the time of first observation/2 is
used.
}
