% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_michaelis_menten_elimination}
\alias{set_michaelis_menten_elimination}
\title{set_michaelis_menten_elimination}
\usage{
set_michaelis_menten_elimination(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Sets elimination to Michaelis-Menten. Initial estimate for CLMM is set to CL and KM is set to
:math:\code{2*max(DV)}.
}
