\name{set_name}
\alias{set_name}
\title{set_name}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.set_name}{Link to Python API reference} (e.g. for correct rendering of equations)


}
\usage{
set_name(model, name)
}
\arguments{

}
\value{

}
