% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_ode_solver}
\alias{set_ode_solver}
\title{set_ode_solver}
\usage{
set_ode_solver(model, solver)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{solver}{(str) Solver to use or NULL for no preference}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Sets ODE solver to use for model

The recognized solvers are CVODS, DGEAR, DVERK, IDA, LSODA and LSODI.
To see the corresponding NONMEM ADVANs see :ref:\code{ADVANS}.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_ode_solver(model, 'LSODA')
}

}
