% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_peripheral_compartments}
\alias{set_peripheral_compartments}
\title{set_peripheral_compartments}
\usage{
set_peripheral_compartments(model, n)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{n}{(int) Number of transit compartments

Model
Reference to same model}
}
\description{
Sets the number of peripheral compartments to a specified number.
}
\seealso{
add_peripheral_compartment

remove_peripheral_compartment
}
