% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_reference_values}
\alias{set_reference_values}
\title{set_reference_values}
\usage{
set_reference_values(model, refs)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{refs}{(list(str=numeric)) Pairs of column names and reference values}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Set reference values for selected columns

All values for each selected column will be replaced. For dose columns
only the values for dosing events will be replaced.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_reference_values(model, list('WGT'=0.5, 'AMT'=4.0))
model$dataset
}

}
