% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_seq_zo_fo_absorption}
\alias{set_seq_zo_fo_absorption}
\title{set_seq_zo_fo_absorption}
\usage{
set_seq_zo_fo_absorption(model)
}
\arguments{
\item{model}{(Model) Model to set or change absorption rate}
}
\value{
(Model) Pharmpy model object
}
\description{
Set or change to sequential zero order first order absorption rate.

Initial estimate for
absorption rate is set the previous rate if available, otherwise it is set to the time of
first observation/2.

Currently lagtime together with sequential zero order first order absorption is not
supported.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_seq_zo_fo_absorption(model)
model$statements$ode_system
}
}
\seealso{
set_instantaneous_absorption

set_zero_order_absorption

set_first_order_absorption
}
