% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_simulation}
\alias{set_simulation}
\title{set_simulation}
\usage{
set_simulation(model, n = 1, seed = 64206)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{n}{(numeric) Number of replicates}

\item{seed}{(numeric) Random seed for the simulation}
}
\value{
(Model) Pharmpy model object
}
\description{
Change model into simulation model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_simulation(model, n=10, seed=1234)
steps <- model$estimation_steps
steps[1]
}

}
