\name{set_transit_compartments}
\alias{set_transit_compartments}
\title{set_transit_compartments}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.set_transit_compartments}{Link to Python API reference} (e.g. for correct rendering of equations)

Set the number of transit compartments of model. Initial estimate for absorption rate is

set the previous rate if available, otherwise it is set to the time of first observation/2

is used.
}
\usage{
set_transit_compartments(model, n)
}
\arguments{

}
\value{

}
