% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_unit}
\alias{set_unit}
\title{set_unit}
\usage{
set_unit(model_or_datainfo, column, unit)
}
\arguments{
\item{model_or_datainfo}{(Model or DataInfo) Model object or DataInfo object}

\item{column}{(str) Name of a column. If the column contains multiple variables, e.g. DV
with multiple DVs, the ID can be specified with a colon. For example "DV:1"
will mean the DV column only when DVID is 1.}

\item{unit}{(str) The unit}
}
\value{
(Model | DataInfo) An updated Model or DataInfo object
}
\description{
Specify the unit of a column

Note that no conversion of units will happen if the unit was already set.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_unit(model, "WGT", "kg")
}
}
\seealso{
convert_unit - Convert between units for a variable
}
