% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_upper_bounds}
\alias{set_upper_bounds}
\title{set_upper_bounds}
\usage{
set_upper_bounds(model, bounds)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{bounds}{(list(str=numeric)) A list of parameter bounds for parameters to change}
}
\value{
(Model) Reference to the same model object
}
\description{
Set parameter upper bounds
}
\examples{
\dontrun{
model <- load_example_model("pheno")
set_upper_bounds(model, list('PTVCL'=10))
model$parameters['PTVCL']
}
}
\seealso{
set_lower_bounds : Set parameter lower bounds

unconstrain_parameters : Remove all constraints of parameters
}
