% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_weibull_absorption}
\alias{set_weibull_absorption}
\title{set_weibull_absorption}
\usage{
set_weibull_absorption(model)
}
\arguments{
\item{model}{(Model) Model to set or change to use Weibull absorption rate}
}
\value{
(Model) Pharmpy model object
}
\description{
Set or change to Weibull type absorption

Initial estimate for absorption rate is set to??

If multiple doses is set to the affected compartment, currently only iv+oral
doses (one of each) is supported

Weibull absorption cannot be used together with lag time and transit compartments.

Assumes that absorption of one does is done when next dose is given.

warning::
This function is still under development.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_weibull_absorption(model)
model$statements$ode_system
}
}
\seealso{
set_zero_order_absorption

set_first_order_absorption
}
