% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_weighted_error_model}
\alias{set_weighted_error_model}
\title{set_weighted_error_model}
\usage{
set_weighted_error_model(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Encode error model with one epsilon and W as weight
}
