% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_zero_order_absorption}
\alias{set_zero_order_absorption}
\title{set_zero_order_absorption}
\usage{
set_zero_order_absorption(model)
}
\arguments{
\item{model}{(Model) Model to set or change to first order absorption rate}
}
\description{
Set or change to zero order absorption rate. Initial estimate for absorption rate is set
the previous rate if available, otherwise it is set to the time of first observation/2 is
used.
}
