% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{simplify_expression}
\alias{simplify_expression}
\title{simplify_expression}
\usage{
simplify_expression(model, expr)
}
\arguments{
\item{model}{(Model) Pharmpy model object}

\item{expr}{(numeric or str) Expression to simplify}
}
\value{
(Expression) Simplified expression
}
\description{
Simplify expression given constraints in model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
simplify_expression(model, "Abs(PTVCL)")
}

}
