% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{solve_ode_system}
\alias{solve_ode_system}
\title{solve_ode_system}
\usage{
solve_ode_system(model)
}
\arguments{
\item{model}{(Model) Pharmpy model object}
}
\value{
(Model) Pharmpy model object
}
\description{
Replace ODE system with analytical solution if possible

Warnings
This function can currently only handle the most simple of ODE systems.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$statements$ode_system
model <- solve_ode_system(model)
}

}
