% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{split_joint_distribution}
\alias{split_joint_distribution}
\title{split_joint_distribution}
\usage{
split_joint_distribution(model, rvs = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{rvs}{(str, vector) Name/names of etas to separate. If NULL, all etas that are IIVs and
non-fixed will become single. NULL is default.}
}
\description{
Splits etas following a joint distribution into separate distributions.
}
