\name{split_rv_block}
\alias{split_rv_block}
\title{split_rv_block}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.split_rv_block}{Link to Python API reference} (e.g. for correct rendering of equations)

Splits an block structure given a list of etas to separate.
}
\usage{
split_rv_block(model, list_of_rvs=None)
}
\arguments{

\item{
model
}{
Model, Pharmpy model to create block effect on.
}
\item{
list_of_rvs
}{
str, list, Name/names of etas to split from block structure. If None, all etas that are IIVs and
}
}
\value{

}
