% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{summarize_errors}
\alias{summarize_errors}
\title{summarize_errors}
\usage{
summarize_errors(context)
}
\arguments{
\item{context}{(Context) Context in which models were run}
}
\value{
(data.frame) A DataFrame of errors with model name, category (error or warning), and an integer as index, an empty DataFrame if there were no errors or warnings found.
}
\description{
Summarize errors and warnings from all runs in a context.

Summarize the errors and warnings found after running the model/models.
}
