% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{summarize_individuals}
\alias{summarize_individuals}
\title{summarize_individuals}
\usage{
summarize_individuals(models)
}
\arguments{
\item{models}{(vector of Models) Input models}
}
\value{
(data.frame | NULL) The summary as a dataframe
}
\description{
Creates a summary dataframe keyed by model-individual pairs for an input
vector of models.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
fit(model)
results <- run_tool(
    model=model,
    mfl='ABSORPTION(ZO);PERIPHERALS(c(1, 2))',
    algorithm='reduced_stepwise'
summarize_individuals([results$start_model, *results$models])
}

}
