% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{summarize_modelfit_results}
\alias{summarize_modelfit_results}
\title{summarize_modelfit_results}
\usage{
summarize_modelfit_results(models)
}
\arguments{
\item{models}{(vector) List of models

pd.DataFrame
A DataFrame of modelfit results, one row per model.}
}
\description{
Summarize results of multiple model runs

Summarize different results after fitting a model, includes runtime, ofv,
and parameter estimates (with errors).
}
