% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{summarize_models}
\alias{summarize_models}
\title{summarize_models}
\usage{
summarize_models(models)
}
\arguments{
\item{models}{(vector) List of models}
}
\description{
Summarize results of multiple model runs, includes runtime, ofv and parameter estimates (with errors).
}
