\name{theta_as_stdev}
\alias{theta_as_stdev}
\title{theta_as_stdev}
\description{
Use thetas to estimate standard deviation of error.

\bold{Usage}

\code{theta_as_stdev(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.theta_as_stdev}{API reference} (for correct rendering of equations, tables etc.).
Use thetas to estimate standard deviation of error

This documentation was automatically generated from Pharmpy (0.26.1).}
