% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{translate_nmtran_time}
\alias{translate_nmtran_time}
\title{translate_nmtran_time}
\usage{
translate_nmtran_time(model)
}
\arguments{
\item{model}{(Model) Pharmpy model object}
}
\value{
(Model) Pharmpy model object
}
\description{
Translate NM-TRAN TIME and DATE column into one TIME column

If dataset of model have special NM-TRAN TIME and DATE columns these
will be translated into one single time column with time in hours.

Warnings
Use this function with caution. For example reset events are currently not taken into account.
}
