% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{unconstrain_parameters}
\alias{unconstrain_parameters}
\title{unconstrain_parameters}
\usage{
unconstrain_parameters(model, parameter_names, strict = TRUE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameter_names}{(array(str)) Remove all constraints for the listed parameters}

\item{strict}{(logical) Whether all parameters in input need to exist in the model. Default is TRUE}
}
\value{
(Model) Pharmpy model object
}
\description{
Remove all constraints from parameters
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$parameters['POP_CL']
model <- unconstrain_parameters(model, c('POP_CL'))
model$parameters['POP_CL']
}
}
\seealso{
set_lower_bounds : Set parameter lower bounds

set_upper_bounds : Set parameter upper bounds

unfix_parameters : Unfix parameters
}
