% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{undrop_columns}
\alias{undrop_columns}
\title{undrop_columns}
\usage{
undrop_columns(model, column_names)
}
\arguments{
\item{model}{(Model) Pharmpy model object}

\item{column_names}{(vector or str) List of column names or one column name to undrop}
}
\value{
(Model) Reference to same model object
}
\description{
Undrop columns of model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
drop_columns(model, c('WGT', 'APGR'), mark=TRUE)
undrop_columns(model, 'WGT')
}
}
\seealso{
drop_dropped_columns : Drop all columns marked as drop

drop_columns : Drop or mark columns as dropped
}
