% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{unload_dataset}
\alias{unload_dataset}
\title{unload_dataset}
\usage{
unload_dataset(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Pharmpy model with dataset removed
}
\description{
Unload the dataset from a model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- unload_dataset(model)
model$dataset is NULL
}

}
