% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{update_initial_individual_estimates}
\alias{update_initial_individual_estimates}
\title{update_initial_individual_estimates}
\usage{
update_initial_individual_estimates(model, individual_estimates, force = TRUE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{individual_estimates}{(array) Individual estimates to use}

\item{force}{(logical) Set to FALSE to only update if the model had initial individual estimates before}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Update initial individual estimates for a model

Updates initial individual estimates for a model.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
results <- load_example_modelfit_results("pheno")
ie <- results$individual_estimates
model <- update_initial_individual_estimates(model, ie)
}

}
