% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{update_inits}
\alias{update_inits}
\title{update_inits}
\usage{
update_inits(model, force_individual_estimates = FALSE)
}
\arguments{
\item{model}{(Model) Pharmpy model to create block effect on.}

\item{force_individual_estimates}{(logical) Whether update of initial individual estimates should be forced.}
}
\description{
Updates initial estimates from previous output. Can be forced if no initial
individual estimates have been read.
}
