% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{update_source}
\alias{update_source}
\title{update_source}
\usage{
update_source(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Update source

Let the code of the underlying source language be updated to reflect
changes in the model object.
}
