% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{use_thetas_for_error_stdev}
\alias{use_thetas_for_error_stdev}
\title{use_thetas_for_error_stdev}
\usage{
use_thetas_for_error_stdev(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Use thetas to estimate standard deviation of error
}
