% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{vpc_plot}
\alias{vpc_plot}
\title{vpc_plot}
\usage{
vpc_plot(
  model,
  simulations,
  binning = "equal_number",
  nbins = 8,
  qi = 0.95,
  ci = 0.95,
  stratify_on = NULL
)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{simulations}{(data.frame) DataFrame containing the simulation data}

\item{binning}{(str) Binning method. Can be "equal_number" or "equal_width". The default is "equal_number".}

\item{nbins}{(numeric) Number of bins. Default is 8.}

\item{qi}{(numeric) Upper quantile. Default is 0.95.}

\item{ci}{(numeric) Confidence interval. Default is 0.95.}

\item{stratify_on}{(str (optional)) Parameter to use for stratification. Optional.}
}
\value{
(alt.Chart) Plot
}
\description{
VPC plot
}
