% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{write_csv}
\alias{write_csv}
\title{write_csv}
\usage{
write_csv(model, path = NULL, force = FALSE)
}
\arguments{
\item{model}{(Model) Model whose dataset to write to file}

\item{path}{(NULL or str or Path) Destination path. Default is to use original path with .csv suffix.}

\item{force}{(logical) Overwrite file with same path. Default is FALSE.}
}
\value{
(Path) path to the written file.
}
\description{
Write dataset to a csv file and updates the datainfo path
}
\examples{
\dontrun{
model <- load_example_model("pheno")
write_csv(model, path="newdataset$csv")
}

}
