% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{write_datainfo}
\alias{write_datainfo}
\title{write_datainfo}
\usage{
write_datainfo(di, path, force = FALSE)
}
\arguments{
\item{di}{(DataInfo) DataInfo object}

\item{path}{(str) Path to write the datainfo file}

\item{force}{(logical) Force overwrite if file already exists}
}
\description{
Write a DataInfo object to a datainfo file
}
