% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{write_dataset}
\alias{write_dataset}
\title{write_dataset}
\usage{
write_dataset(model, path = NULL, force = FALSE, type = "csv")
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{path}{(str (optional)) Destination path. Default is to use original path with .csv suffix.}

\item{force}{(logical) Overwrite file with same path. Default is FALSE.}

\item{type}{(str) Can only be csv}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Write dataset to file and updates the datainfo path

Curretly supports csv files.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- write_dataset(model, path="newdataset$csv")
}

}
