\name{write_model}
\alias{write_model}
\title{write_model}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.write_model}{Link to Python API reference} (e.g. for correct rendering of equations)

Write model to file
}
\usage{
write_model(model, path='', force=False)
}
\arguments{

}
\value{

}
