\name{zero_order_absorption}
\alias{zero_order_absorption}
\title{zero_order_absorption}
\description{
Set or change to zero order absorption rate.

\bold{Usage}

\code{zero_order_absorption(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Set or change to zero order absorption rate. Initial estimate for absorption rate is set

the previous rate if available, otherwise it is set to the time of first observation/2 is

used.
}
\arguments{

\item{
model
}{
Model, Model to set or change to first order absorption rate
}
}
