\name{zero_order_elimination}
\alias{zero_order_elimination}
\title{zero_order_elimination}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.zero_order_elimination}{Link to Python API reference} (e.g. for correct rendering of equations)

Sets elimination to zero order. Initial estimate for KM is set to 1% of smallest

observation.
}
\usage{
zero_order_elimination(model)
}
\arguments{

}
\value{

}
