% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_effect_compartment}
\alias{add_effect_compartment}
\title{add_effect_compartment}
\usage{
add_effect_compartment(model, expr)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{expr}{(str) Name of the PD effect function. Valid names are: baseline, linear, Emax, sigmoid, step, loglin}
}
\value{
(Model) Pharmpy model object
}
\description{
Add an effect compartment.

Implemented PD models are:
\itemize{
\item Baseline:
(equation could not be rendered, see API doc on website)
\item Linear:
(equation could not be rendered, see API doc on website)
\item Emax:
(equation could not be rendered, see API doc on website)
\item Step effect:
(equation could not be rendered, see API doc on website)
\item Sigmoidal:
(equation could not be rendered, see API doc on website)
\item Log-linear:
(equation could not be rendered, see API doc on website)
}
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- add_effect_compartment(model, "linear")
model$statements$ode_system$find_compartment("EFFECT")
}

}
