% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_individual_parameter}
\alias{add_individual_parameter}
\title{add_individual_parameter}
\usage{
add_individual_parameter(model, name, init = 0.1, lower = 0)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{name}{(str) Name of individual/pk parameter}

\item{init}{(numeric) Initial estimate of the population parameter}

\item{lower}{(numeric) Lower bound for the population parameter}
}
\value{
(Model) Pharmpy model object
}
\description{
Add an individual or pk parameter to a model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- add_individual_parameter(model, "KA")
model$statements$find_assignment("KA")
}

}
