\name{add_lag_time}
\alias{add_lag_time}
\title{add_lag_time}
\description{
Add lag time to the dose compartment of model.

\bold{Usage}

\code{add_lag_time(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Add lag time to the dose compartment of model. Initial estimate for lag time is set the

previous lag time if available, otherwise it is set to the time of first observation/2 is

used.

This documentation was automatically generated from Pharmpy (0.22.0).}
