% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{add_metabolite}
\alias{add_metabolite}
\title{add_metabolite}
\usage{
add_metabolite(model, drug_dvid = 1, presystemic = FALSE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{drug_dvid}{(numeric) DVID for drug (assuming all other DVIDs being for metabolites)}

\item{presystemic}{(logical) Decide wether or not to add metabolite as a presystemetic fixed drug.}
}
\value{
(Model) Pharmpy model object
}
\description{
Adds a metabolite compartment to a model

The flow from the central compartment to the metabolite compartment
will be unidirectional.

Presystemic indicate that the metabolite compartment will be
directly connected to the DEPOT. If a depot compartment is not present,
one will be created.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- add_metabolite(model)
}

}
