\name{add_parameter}
\alias{add_parameter}
\title{add_parameter}
\description{
Add an individual or pk parameter to a model.

\bold{Usage}

\code{add_parameter(model, name)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.add_parameter}{API reference} (for correct rendering of equations, tables etc.).
Add an individual or pk parameter to a model

This documentation was automatically generated from Pharmpy (0.26.1).}
