\name{additive_error}
\alias{additive_error}
\title{additive_error}
\description{
Set an additive error model.

\bold{Usage}

\code{additive_error(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Set an additive error model. Initial estimate for new sigma is :math:`(min(DV)/2)²`.
}
\arguments{

}
