\name{bolus_absorption}
\alias{bolus_absorption}
\title{bolus_absorption}
\description{
Set or change to bolus absorption rate..

\bold{Usage}

\code{bolus_absorption(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.bolus_absorption}{API reference} (for correct rendering of equations, tables etc.).
Set or change to bolus absorption rate.

This documentation was automatically generated from Pharmpy (0.23.4).}
\arguments{

\item{
model
}{
Model. Model to set or change absorption rate
}
}
