% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{bump_model_number}
\alias{bump_model_number}
\title{bump_model_number}
\usage{
bump_model_number(model, path = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model object}

\item{path}{(Path in which to find next unique number) Default is to not look for files.}
}
\value{
(Model) Reference to the same model object
}
\description{
If the model name ends in a number increase it

If path is set increase the number until no file exists
with the same name in path.
If model name does not end in a number do nothing.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$name <- "run2"
bump_model_number(model)
model$name
}

}
