% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_aic}
\alias{calculate_aic}
\title{calculate_aic}
\usage{
calculate_aic(model)
}
\arguments{
\item{model}{(Model) Pharmpy model object}
}
\value{
(numeric) AIC of model fit
}
\description{
Calculate final AIC for model assuming the OFV to be -2LL

AIC = OFV + 2*n_estimated_parameters
}
