% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_bic}
\alias{calculate_bic}
\title{calculate_bic}
\usage{
calculate_bic(model)
}
\arguments{
\item{model}{(Model) Pharmpy model object}
}
\value{
(numeric) BIC of model fit
}
\description{
Calculate final BIC value assuming the OFV to be -2LL

BIC = OFV + n_estimated_parameters * log(n_observations)
}
