% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_corr_from_cov}
\alias{calculate_corr_from_cov}
\title{calculate_corr_from_cov}
\usage{
calculate_corr_from_cov(cov)
}
\arguments{
\item{cov}{(data.frame) Covariance matrix}
}
\value{
(data.frame) Correlation matrix
}
\description{
Calculate correlation matrix from a covariance matrix
}
\examples{
\dontrun{
model <- load_example_model("pheno")
cov <- model$modelfit_results$covariance_matrix
cov
calculate_corr_from_cov(cov)
}
}
\seealso{
calculate_se_from_cov : Standard errors from covariance matrix

calculate_se_from_inf : Standard errors from information matrix

calculate_cov_from_inf : Covariance matrix from information matrix

calculate_cov_from_corrse : Covariance matrix from correlation matrix and standard errors

calculate_inf_from_cov : Information matrix from covariance matrix

calculate_inf_from_corrse : Information matrix from correlation matrix and standard errors

calculate_corr_from_inf : Correlation matrix from information matrix
}
