% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_epsilon_gradient_expression}
\alias{calculate_epsilon_gradient_expression}
\title{calculate_epsilon_gradient_expression}
\usage{
calculate_epsilon_gradient_expression(model)
}
\arguments{
\item{model}{(Model) Pharmpy model object}
}
\value{
(Expression) Symbolic expression
}
\description{
Calculate the symbolic expression for the epsilon gradient

This function currently only support models without ODE systems
}
\examples{
\dontrun{
model <- load_example_model("pheno_linear")
calculate_epsilon_gradient_expression(model)
}
}
\seealso{
calculate_eta_gradient_expression : Eta gradient
}
