% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_eta_shrinkage}
\alias{calculate_eta_shrinkage}
\title{calculate_eta_shrinkage}
\usage{
calculate_eta_shrinkage(model, sd = FALSE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{sd}{(logical) Calculate shrinkage on the standard deviation scale (default is to calculate on the
variance scale)

Series
Shrinkage for each eta}
}
\description{
Calculate eta shrinkage for each eta
}
\seealso{
calculate_individual_shrinkage
}
