% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_individual_shrinkage}
\alias{calculate_individual_shrinkage}
\title{calculate_individual_shrinkage}
\usage{
calculate_individual_shrinkage(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Calculate the individual eta-shrinkage

Definition: ieta_shr = (var(eta) / omega)
}
