% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_individual_shrinkage}
\alias{calculate_individual_shrinkage}
\title{calculate_individual_shrinkage}
\usage{
calculate_individual_shrinkage(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(DataFrame) Shrinkage for each eta and individual
}
\description{
Calculate the individual eta-shrinkage

Definition: ieta_shr = (var(eta) / omega)
}
\examples{
\dontrun{
model <- load_example_model("pheno")
calculate_individual_shrinkage(model)
}
}
\seealso{
calculate_eta_shrinkage
}
