% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{calculate_pk_parameters_statistics}
\alias{calculate_pk_parameters_statistics}
\title{calculate_pk_parameters_statistics}
\usage{
calculate_pk_parameters_statistics(model, seed = NULL)
}
\arguments{
\item{model}{(Model) A previously estimated model}

\item{seed}{(int or numpy rng)

pd.DataFrame
A DataFrame of statistics indexed on parameter and covariate value.}
}
\description{
Calculate statistics for common pharmacokinetic parameters

Calculate the mean (expected value of the distribution), variance
(variance of the distribution) and standard error for some individual
pre-defined pharmacokinetic parameters.
}
\seealso{
calculate_individual_parameter_statistics : Calculation of statistics for arbitrary parameters
}
