% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pharmpy.R
\name{check_pharmpy}
\alias{check_pharmpy}
\title{Checks version of Pharmpy/pharmr}
\usage{
check_pharmpy(pharmpy_version)
}
\arguments{
\item{pharmpy_version}{(str) version number as string}
}
\description{
Checks whether Pharmpy and pharmr has the same version
}
