\name{combined_error}
\alias{combined_error}
\title{combined_error}
\description{
Set a combined error model.

\bold{Usage}

\code{combined_error(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Set a combined error model. Initial estimates for new sigmas are :math:`(min(DV)/2)²` for

proportional and 0.09 for additive.
}
\arguments{

}
