% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{convert_model}
\alias{convert_model}
\title{convert_model}
\usage{
convert_model(model, to_format)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{to_format}{(str) Name of format to convert into. Currently supported 'generic', 'nlmixr', 'nonmem', and 'rxode'}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Convert model to other format

Note that the operation is not done inplace.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
converted_model <- convert_model(model, "nlmixr")
}

}
