% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{convert_model}
\alias{convert_model}
\title{convert_model}
\usage{
convert_model(model, to_format)
}
\arguments{
\item{model}{(Model) Model to convert}

\item{to_format}{(str) Name of format to convert into. Currently supported 'nlmixr' and 'nonmem'}
}
\value{
(Model) New model object with new underlying model format
}
\description{
Convert model to other format
}
\examples{
\dontrun{
model <- load_example_model("pheno")
converted_model <- convert_model(model, "nlmixr")
}

}
