\name{convert_model}
\alias{convert_model}
\title{convert_model}
\description{
Convert model to other format.

\bold{Usage}

\code{convert_model(model, to)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.convert_model}{API reference} (for correct rendering of equations, tables etc.).
Convert model to other format

This documentation was automatically generated from Pharmpy (0.26.1).}
\arguments{

\item{
model
}{
Model. Model to convert
}
\item{
to
}{
character. Name of format to convert into. Currently supported 'nlmixr' Results ------- Model New model object with new underlying model format.
}
}
