% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{copy_model}
\alias{copy_model}
\title{copy_model}
\usage{
copy_model(model, name = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{name}{(str) Optional new name of model

Model
A copy of the input model}
}
\description{
Copies model to a new model object
}
