\name{copy_model}
\alias{copy_model}
\title{copy_model}
\description{
NULL.

\bold{Usage}

\code{copy_model(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.copy_model}{API reference} (for correct rendering of equations, tables etc.).


This documentation was automatically generated from Pharmpy (0.22.0).}
