\name{copy_model}
\alias{copy_model}
\title{copy_model}
\description{
Copies model to a new model object.

\bold{Usage}

\code{copy_model(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.copy_model}{API reference} (for correct rendering of equations, tables etc.).
Copies model to a new model object

This documentation was automatically generated from Pharmpy (0.26.1).}
