% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{copy_model}
\alias{copy_model}
\title{copy_model}
\usage{
copy_model(model, name = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{name}{(str (optional)) Optional new name of model}
}
\value{
(Model) A copy of the input model
}
\description{
Copies model to a new model object
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model_copy <- copy_model(model, "pheno2")
}

}
