% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{create_results}
\alias{create_results}
\title{create_results}
\usage{
create_results(path)
}
\arguments{
\item{path}{(str, Path) Path to run directory
kwargs
Arguments to pass to tool specific create results function

Results object for tool}
}
\description{
Create/recalculate results object given path to run directory
}
\seealso{
read_results
}
