\name{create_results}
\alias{create_results}
\title{create_results}
\description{
NULL.

\bold{Usage}

\code{create_results(path, **kwargs)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).

}
