\name{create_results}
\alias{create_results}
\title{create_results}
\description{
NA.

\bold{Usage}

\code{create_results(path, **kwargs)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.create_results}{API reference} (for correct rendering of equations, tables etc.).


This documentation was automatically generated from Pharmpy (0.25.1).}
