% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{expand_additional_doses}
\alias{expand_additional_doses}
\title{expand_additional_doses}
\usage{
expand_additional_doses(model, flag = FALSE)
}
\arguments{
\item{model}{(Model) Pharmpy model object}

\item{flag}{(logical) TRUE to add a boolean EXPANDED column to mark added records. In this case all
columns in the original dataset will be kept. Care needs to be taken to handle
the new dataset.}
}
\value{
(Model) Pharmpy model object
}
\description{
Expand additional doses into separate dose records
}
