\name{explicit_odes}
\alias{explicit_odes}
\title{explicit_odes}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.explicit_odes}{Link to Python API reference} (e.g. for correct rendering of equations)

Convert model from compartmental system to explicit ODE system

or do nothing if it already has an explicit ODE system
}
\usage{
explicit_odes(model)
}
\arguments{

}
\value{

}
