\name{explicit_odes}
\alias{explicit_odes}
\title{explicit_odes}
\description{
Convert model from compartmental system to explicit ODE system.

\bold{Usage}

\code{explicit_odes(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.explicit_odes}{API reference} (for correct rendering of equations, tables etc.).
Convert model from compartmental system to explicit ODE system

or do nothing if it already has an explicit ODE system

This documentation was automatically generated from Pharmpy (0.22.0).}
