\name{first_order_absorption}
\alias{first_order_absorption}
\title{first_order_absorption}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.first_order_absorption}{Link to Python API reference} (e.g. for correct rendering of equations)

Set or change to first order absorption rate. Initial estimate for absorption rate is set

the previous rate if available, otherwise it is set to the time of first observation/2 is used.
}
\usage{
first_order_absorption(model)
}
\arguments{

}
\value{

}
