\name{first_order_elimination}
\alias{first_order_elimination}
\title{first_order_elimination}
\description{
NA.

\bold{Usage}

\code{first_order_elimination(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.first_order_elimination}{API reference} (for correct rendering of equations, tables etc.).


This documentation was automatically generated from Pharmpy (0.23.3).}
