% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{fit}
\alias{fit}
\title{fit}
\usage{
fit(model_or_models, tool = NULL, path = NULL)
}
\arguments{
\item{model_or_models}{(Model or array(Model)) List of models or one single model}

\item{tool}{(str (optional)) Estimation tool to use. NULL to use default}

\item{path}{(str (optional)) Path to fit directory}
}
\value{
(ModelfitResults | vector of ModelfitResults) ModelfitResults for the model or models
}
\description{
Fit models.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
results <- fit(model)
}
}
\seealso{
run_tool
}
