% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{fit}
\alias{fit}
\title{fit}
\usage{
fit(models)
}
\arguments{
\item{models}{(vector) List of models or one single model

Model
Reference to same model}
}
\description{
Fit models.
}
\seealso{
run_tool
}
