% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{fit}
\alias{fit}
\title{fit}
\usage{
fit(models)
}
\arguments{
\item{models}{(vector) List of models or one single model}
}
\value{
(Model) Reference to same model
}
\description{
Fit models.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
fit(model)
}
}
\seealso{
run_tool
}
