% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{fit}
\alias{fit}
\title{fit}
\usage{
fit(model_or_models, tool = NULL)
}
\arguments{
\item{model_or_models}{(Model | vector of Models) List of models or one single model}

\item{tool}{(str) Estimation tool to use. NULL to use default}
}
\value{
(Model | vector of Models) Input model or models with model fit results
}
\description{
Fit models.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
fit(model)
}
}
\seealso{
run_tool
}
