\name{fit}
\alias{fit}
\title{fit}
\description{
NA.

\bold{Usage}

\code{fit(models)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.fit}{API reference} (for correct rendering of equations, tables etc.).


This documentation was automatically generated from Pharmpy (0.26.1).}
