% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{fix_parameters}
\alias{fix_parameters}
\title{fix_parameters}
\usage{
fix_parameters(model, parameter_names)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameter_names}{(vector or str) one parameter name or a vector of parameter names

Model
Reference to the same model object}
}
\description{
Fix parameters

Fix all listed parameters
}
\seealso{
fix_parameters_to : Fixing and setting parameter initial estimates in the same function

unfix_paramaters : Unfixing parameters

unfix_paramaters_to : Unfixing parameters and setting a new initial estimate in the same

function
}
