\name{fix_parameters}
\alias{fix_parameters}
\title{fix_parameters}
\description{
Fix parameters.

\bold{Usage}

\code{fix_parameters(model, parameter_names)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.fix_parameters}{API reference} (for correct rendering of equations, tables etc.).
Fix parameters

Fix all listed parameters

This documentation was automatically generated from Pharmpy (0.22.0).}
\arguments{

\item{
model
}{
Model
}
\item{
parameter_names
}{
vector or character. one parameter name or a list of parameter names
}
}
\value{
model : Model
}
