% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{fix_parameters}
\alias{fix_parameters}
\title{fix_parameters}
\usage{
fix_parameters(model, parameter_names)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameter_names}{(vector or str) one parameter name or a vector of parameter names}
}
\value{
Model (Reference to the same model)
}
\description{
Fix parameters

Fix all listed parameters
}
