\name{fix_parameters}
\alias{fix_parameters}
\title{fix_parameters}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.fix_parameters}{Link to Python API reference} (e.g. for correct rendering of equations)

Fix parameters

Fix all listed parameters
}
\usage{
fix_parameters(model, parameter_names)
}
\arguments{

}
\value{
model: Model
}
