\name{fix_parameters}
\alias{fix_parameters}
\title{fix_parameters}
\description{
Fix parameters.

\bold{Usage}

\code{fix_parameters(model, parameter_names)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Fix parameters

Fix all listed parameters
}
\arguments{

\item{
model
}{
Model, parameter_names : list or str
}
\item{
parameter_names
}{
list or str, one parameter name or a list of parameter names
}
}
\value{
model : Model
}
