% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{fix_parameters}
\alias{fix_parameters}
\title{fix_parameters}
\usage{
fix_parameters(model, parameter_names, strict = TRUE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameter_names}{(array(str) or str) one parameter name or a vector of parameter names}

\item{strict}{(logical) Whether all parameters in input need to exist in the model. Default is TRUE}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Fix parameters

Fix all listed parameters
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$parameters['POP_CL']
model <- fix_parameters(model, 'POP_CL')
model$parameters['POP_CL']
}
}
\seealso{
fix_or_unfix_parameters : Fix or unfix parameters (given boolean)

fix_parameters_to : Fixing and setting parameter initial estimates in the same function

unfix_paramaters : Unfixing parameters

unfix_paramaters_to : Unfixing parameters and setting a new initial estimate in the same

function
}
