% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{fix_parameters_to}
\alias{fix_parameters_to}
\title{fix_parameters_to}
\usage{
fix_parameters_to(model, parameter_names, values)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameter_names}{(vector or str) one parameter name or a vector of parameter names}

\item{values}{(vector or float) one value or a vector of values (must be equal to number of parameter_names)}
}
\value{
Model (Reference to the same model object)
}
\description{
Fix parameters to

Fix all listed parameters to specified value/values
}
