% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{fix_parameters_to}
\alias{fix_parameters_to}
\title{fix_parameters_to}
\usage{
fix_parameters_to(model, inits, strict = TRUE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{inits}{(list(str=numeric)) Inits for all parameters to fix and set init}

\item{strict}{(logical) Whether all parameters in input need to exist in the model. Default is TRUE}
}
\value{
(Model) Pharmpy model object
}
\description{
Fix parameters to

Fix all listed parameters to specified value/values
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$parameters['POP_CL']
model <- fix_parameters_to(model, list('POP_CL'=0.5))
model$parameters['POP_CL']
}
}
\seealso{
fix_parameters : Fix parameters

fix_or_unfix_parameters : Fix or unfix parameters (given boolean)

unfix_paramaters : Unfixing parameters

unfix_paramaters_to : Unfixing parameters and setting a new initial estimate in the same

function
}
