\name{fix_parameters_to}
\alias{fix_parameters_to}
\title{fix_parameters_to}
\description{
Fix parameters to.

\bold{Usage}

\code{fix_parameters_to(model, parameter_names, values)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.fix_parameters_to}{API reference} (for correct rendering of equations, tables etc.).
Fix parameters to

Fix all listed parameters to specified value/values

This documentation was automatically generated from Pharmpy (0.22.0).}
\arguments{

\item{
model
}{
Model
}
\item{
parameter_names
}{
vector or character. one parameter name or a list of parameter names
}
\item{
values
}{
vector or numeric. one value or a list of values (must be equal to number of parameter_names)
}
}
\value{
model : Model
}
