% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_admid}
\alias{get_admid}
\title{get_admid}
\usage{
get_admid(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(data.frame) ADMID
}
\description{
Get the admid from model dataset

If an administration column is present this will be extracted otherwise
an admid column will be created based on the admids of the present doses.
This is dependent on the presence of a CMT column to be generated correctly.

When generated, admids of events in between doses is set to the last used
admid.
}
