% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_central_volume_and_clearance}
\alias{get_central_volume_and_clearance}
\title{get_central_volume_and_clearance}
\usage{
get_central_volume_and_clearance(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(sympy.Symbol) Volume symbol sympy.Symbol Clearance symbol
}
\description{
Get the volume and clearance parameters
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_central_volume_and_clearance(model)
}

}
