% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_lag_times}
\alias{get_lag_times}
\title{get_lag_times}
\usage{
get_lag_times(model)
}
\arguments{
\item{model}{(Pharmpy model)
Result
list
Dictionary from compartment name to lag time expression}
}
\description{
Get lag times for all compartments
}
