% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_number_of_individuals}
\alias{get_number_of_individuals}
\title{get_number_of_individuals}
\usage{
get_number_of_individuals(model)
}
\arguments{
\item{model}{(Model) Pharmpy model

int
Number of individuals in the model dataset}
}
\description{
Retrieve the number of individuals in the model dataset
}
