% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_number_of_observations}
\alias{get_number_of_observations}
\title{get_number_of_observations}
\usage{
get_number_of_observations(model)
}
\arguments{
\item{model}{(Model) Pharmpy model

int
Number of observations in the model dataset}
}
\description{
Retrieve the total number of observations in the model dataset
}
\note{
For NONMEM models this is the number of observations of the active dataset, i.e. after filteringof IGNORE and ACCEPT and removal of individuals with no observations.
}
\seealso{
get_number_of_individuals : Get the number of individuals in a dataset

get_number_of_observations_per_individual : Get the number of observations per individual in a

dataset
}
