% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_number_of_observations}
\alias{get_number_of_observations}
\title{get_number_of_observations}
\usage{
get_number_of_observations(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Retrieve the total number of observations in the model dataset
}
