% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_number_of_observations_per_individual}
\alias{get_number_of_observations_per_individual}
\title{get_number_of_observations_per_individual}
\usage{
get_number_of_observations_per_individual(model)
}
\arguments{
\item{model}{(Model) Pharmpy model

pd.Series
Number of observations in the model dataset}
}
\description{
Number of observations for each individual
}
\note{
For NONMEM models this is the individuals and number of observations of the active dataset, i.e.after filtering of IGNORE and ACCEPT and removal of individuals with no observations.
}
\seealso{
get_number_of_individuals : Get the number of individuals in a dataset

get_number_of_observations_per_individual : Get the number of observations per individual in a

dataset
}
