% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_observations}
\alias{get_observations}
\title{get_observations}
\usage{
get_observations(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(data.frame) Observations indexed over ID and TIME
}
\description{
Get observations from dataset
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_observations(model)
}
}
\seealso{
get_number_of_observations

get_number_of_observations_per_individual
}
