% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{get_omegas}
\alias{get_omegas}
\title{get_omegas}
\usage{
get_omegas(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Parameters) A copy of all omega parameters
}
\description{
Get all omegas (variability parameters) of a model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
get_omegas(model)
}
}
\seealso{
get_thetas : Get theta parameters

get_sigmas : Get sigma parameters
}
