% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{greekify_model}
\alias{greekify_model}
\title{greekify_model}
\usage{
greekify_model(model, named_subscripts = FALSE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{named_subscripts}{(logical) Use previous parameter names as subscripts. Default is to use integer subscripts

Result
Model
Reference to the same model}
}
\description{
Convert to using greek letters for all population parameters
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$statements
greekify_model(cleanup_model(model))
model$statements
}

}
