\name{has_additive_error}
\alias{has_additive_error}
\title{has_additive_error}
\description{
Check if a model has an additive error model.

\bold{Usage}

\code{has_additive_error(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Check if a model has an additive error model
}
\arguments{

}
