% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_additive_error_model}
\alias{has_additive_error_model}
\title{has_additive_error_model}
\usage{
has_additive_error_model(model, dv = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{dv}{(Expr or str or numeric (optional)) Name or DVID of dependent variable. NULL for the default (first or only)}
}
\value{
(logical) TRUE if the model has an additive error model and FALSE otherwise
}
\description{
Check if a model has an additive error model

Multiple dependent variables are supported. By default the only (in case of one) or the
first (in case of many) dependent variable is going to be checked.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
has_additive_error_model(model)
}
}
\seealso{
has_proportional_error_model : Check if a model has a proportional error model

has_combined_error_model : Check if a model has a combined error model

has_weighted_error_model : Check if a model has a weighted error model
}
