\name{has_combined_error}
\alias{has_combined_error}
\title{has_combined_error}
\description{
Check if a model has a combined additive and proportinal error model.

\bold{Usage}

\code{has_combined_error(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.has_combined_error}{API reference} (for correct rendering of equations, tables etc.).
Check if a model has a combined additive and proportinal error model

This documentation was automatically generated from Pharmpy (0.24.2).}
\arguments{

\item{
model
}{
Model. The model to check
}
}
