% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_combined_error_model}
\alias{has_combined_error_model}
\title{has_combined_error_model}
\usage{
has_combined_error_model(model)
}
\arguments{
\item{model}{(Model) The model to check}
}
\value{
(logical) TRUE if the model has a combined error model and FALSE otherwise
}
\description{
Check if a model has a combined additive and proportinal error model
}
\examples{
\dontrun{
model <- load_example_model("pheno")
has_combined_error_model(model)
}
}
\seealso{
has_additive_error_model : Check if a model has an additive error model

has_proportional_error_model : Check if a model has a proportional error model
}
