\name{has_proportional_error}
\alias{has_proportional_error}
\title{has_proportional_error}
\description{
Check if a model has a proportional error model.

\bold{Usage}

\code{has_proportional_error(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.has_proportional_error}{API reference} (for correct rendering of equations, tables etc.).
Check if a model has a proportional error model

This documentation was automatically generated from Pharmpy (0.25.0).}
\arguments{

\item{
model
}{
Model. The model to check
}
}
