% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_proportional_error_model}
\alias{has_proportional_error_model}
\title{has_proportional_error_model}
\usage{
has_proportional_error_model(model)
}
\arguments{
\item{model}{(Model) The model to check

logical
TRUE if the model has a proportional error model and FALSE otherwise}
}
\description{
Check if a model has a proportional error model
}
\seealso{
has_additive_error_model : Check if a model has an additive error model

has_combined_error_model : Check if a model has a combined error model
}
