% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{has_zero_order_absorption}
\alias{has_zero_order_absorption}
\title{has_zero_order_absorption}
\usage{
has_zero_order_absorption(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Check if ode system describes a zero order absorption

currently defined as having Infusion dose with rate not in dataset
}
