\name{michaelis_menten_elimination}
\alias{michaelis_menten_elimination}
\title{michaelis_menten_elimination}
\description{
Sets elimination to Michaelis-Menten.

\bold{Usage}

\code{michaelis_menten_elimination(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.michaelis_menten_elimination}{API reference} (for correct rendering of equations, tables etc.).
Sets elimination to Michaelis-Menten. Initial estimate for CLMM is set to CL and KM is set to

:math:`2*max(DV)`.

This documentation was automatically generated from Pharmpy (0.22.10).}
