\name{mixed_mm_fo_elimination}
\alias{mixed_mm_fo_elimination}
\title{mixed_mm_fo_elimination}
\description{
Sets elimination to mixed Michaelis-Menten and first order.

\bold{Usage}

\code{mixed_mm_fo_elimination(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.mixed_mm_fo_elimination}{API reference} (for correct rendering of equations, tables etc.).
Sets elimination to mixed Michaelis-Menten and first order. Initial estimate for CLMM is set

to CL/2 and KM is set to :math:`2*max(DV)`.

This documentation was automatically generated from Pharmpy (0.25.1).}
