% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{plot_abs_cwres_vs_ipred}
\alias{plot_abs_cwres_vs_ipred}
\title{plot_abs_cwres_vs_ipred}
\usage{
plot_abs_cwres_vs_ipred(model, predictions, residuals)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{predictions}{(data.frame) DataFrame containing the predictions}

\item{residuals}{(data.frame) DataFrame containing the residuals}
}
\value{
(alt.Chart) Plot
}
\description{
Plot \|CWRES\| vs IPRED
}
\examples{
\dontrun{
}

}
