% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{plot_cwres_vs_idv}
\alias{plot_cwres_vs_idv}
\title{plot_cwres_vs_idv}
\usage{
plot_cwres_vs_idv(model, residuals)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{residuals}{(data.frame) DataFrame containing CWRES}
}
\value{
(alt.Chart) Plot
}
\description{
Plot CWRES vs idv
}
