% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{plot_dv_vs_ipred}
\alias{plot_dv_vs_ipred}
\title{plot_dv_vs_ipred}
\usage{
plot_dv_vs_ipred(model, predictions, strat = NULL, bins = 8)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{predictions}{(data.frame) DataFrame containing the predictions}

\item{strat}{(str) Name of parameter for stratification}

\item{bins}{(numeric) Number of bins for stratification}
}
\value{
(alt.Chart) Plot
}
\description{
Plot DV vs IPRED
}
\examples{
\dontrun{
}

}
