% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{plot_individual_predictions}
\alias{plot_individual_predictions}
\title{plot_individual_predictions}
\usage{
plot_individual_predictions(model, predictions, individuals = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{predictions}{(data.frame) One column for each type of prediction}

\item{individuals}{(array(numeric) (optional)) A vector of individuals to include. NULL for all individuals}
}
\value{
(alt.Chart) Plot
}
\description{
Plot DV and predictions grouped on individuals
}
\examples{
\dontrun{
model <- load_example_model("pheno")
res <- load_example_modelfit_results("pheno")
plot_individual_predictions(model, res$predictions, individuals=c(1, 2, 3, 4, 5))
}

}
