% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{plot_iofv_vs_iofv}
\alias{plot_iofv_vs_iofv}
\title{plot_iofv_vs_iofv}
\usage{
plot_iofv_vs_iofv(model, other)
}
\arguments{
\item{model}{(Model) The first model}

\item{other}{(Model) The second model}
}
\value{
(alt.Chart) Scatterplot
}
\description{
Plot individual OFV of two models against each other
}
