% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{predict_influential_individuals}
\alias{predict_influential_individuals}
\title{predict_influential_individuals}
\usage{
predict_influential_individuals(model, results, cutoff = 3.84)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{results}{(ModelfitResults) Results for model}

\item{cutoff}{(numeric) Cutoff threshold for a dofv signalling an influential individual}
}
\value{
(data.frame) Dataframe over the individuals with a \code{dofv} column containing the raw predicted delta-OFV and an \code{influential} column with a boolean to tell whether the individual is influential or not.
}
\description{
Predict influential individuals for a model using a machine learning model.
}
\seealso{
predict_influential_outliers

predict_outliers
}
