% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{predict_influential_outliers}
\alias{predict_influential_outliers}
\title{predict_influential_outliers}
\usage{
predict_influential_outliers(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(pd.Dataframe) Dataframe over the individuals with a \code{outliers} and \code{dofv} columns containing the raw predictions and \code{influential}, \code{outlier} and \code{influential_outlier} boolean columns.
}
\description{
Predict influential outliers for a model using a machine learning model.
}
\seealso{
predict_influential_individuals

predict_outliers
}
