% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{print_fit_summary}
\alias{print_fit_summary}
\title{print_fit_summary}
\usage{
print_fit_summary(model, modelfit_results)
}
\arguments{
\item{model}{(Model) Pharmpy model object}

\item{modelfit_results}{(ModelfitResults) Pharmpy ModelfitResults object}
}
\description{
Print a summary of the model fit
}
