\name{proportional_error}
\alias{proportional_error}
\title{proportional_error}
\description{
Set a proportional error model.

\bold{Usage}

\code{proportional_error(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Set a proportional error model. Initial estimate for new sigma is 0.09.
}
\arguments{

\item{
model
}{
Model, Set error model for this model
}
}
