% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{rank_models}
\alias{rank_models}
\title{rank_models}
\usage{
rank_models(
  base_model,
  base_model_res,
  models,
  models_res,
  parent_dict = NULL,
  strictness = "minimization_successful",
  rank_type = "ofv",
  cutoff = NULL,
  bic_type = "mixed",
  ...
)
}
\arguments{
\item{base_model}{(Model) Base model to compare to}

\item{base_model_res}{(ModelfitResults) Results of base model}

\item{models}{(array(Model)) List of models}

\item{models_res}{(array(ModelfitResults)) List of modelfit results}

\item{parent_dict}{(list(str=str) or list(Model=Model) (optional)) Dict where key is child and value is parent. Only relevant for LRT, if NULL base will be set as parent}

\item{strictness}{(str (optional)) Strictness criteria that are allowed for ranking. Default is "minimization_successful".}

\item{rank_type}{(str) Name of ranking type. Available options are 'ofv', 'aic', 'bic', 'lrt' (OFV with LRT)}

\item{cutoff}{(numeric (optional)) Value to use as cutoff. If using LRT, cutoff denotes p-value. Default is NULL}

\item{bic_type}{(str) Type of BIC to calculate. Default is the mixed effects.}

\item{...}{Arguments to pass to calculate BIC (such as \code{mult_test_p} and \code{mult_test_p})}
}
\value{
(data.frame) DataFrame of the ranked models
}
\description{
Ranks a vector of models

Ranks a vector of models with a given ranking function
}
\examples{
\dontrun{
model_1 <- load_example_model("pheno")
model_2 <- load_example_model("pheno_linear")
rank_models(model_1, c(model_2),
 rank_type='lrt')
}

}
