% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{read_model}
\alias{read_model}
\title{read_model}
\usage{
read_model(path)
}
\arguments{
\item{path}{(str or Path) Path to model}
}
\value{
(Model) Read model object
}
\description{
Read model from file
}
\examples{
\dontrun{
read_model("/home/run1$mod")
}
}
\seealso{
read_model_from_string : Read model from string
}
