\name{read_model}
\alias{read_model}
\title{read_model}
\description{
Read model from file.

\bold{Usage}

\code{read_model(path)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.read_model}{API reference} (for correct rendering of equations, tables etc.).
Read model from file

This documentation was automatically generated from Pharmpy (0.22.0).}
