% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{read_model_from_string}
\alias{read_model_from_string}
\title{read_model_from_string}
\usage{
read_model_from_string(code)
}
\arguments{
\item{code}{(str) Model code to read

Model
Read model object}
}
\description{
Read model from the model code in a string
}
