\name{read_model_from_string}
\alias{read_model_from_string}
\title{read_model_from_string}
\description{
Read model directly from the model code in a string.

\bold{Usage}

\code{read_model_from_string(code)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.read_model_from_string}{API reference} (for correct rendering of equations, tables etc.).
Read model directly from the model code in a string

This documentation was automatically generated from Pharmpy (0.23.3).}
