% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{read_model_from_string}
\alias{read_model_from_string}
\title{read_model_from_string}
\usage{
read_model_from_string(code)
}
\arguments{
\item{code}{(str) Model code to read}
}
\value{
(Model) Pharmpy model
}
\description{
Read model from the model code in a string
}
\examples{
\dontrun{
s <- "$PROBLEM
$INPUT ID DV TIME
$DATA file$csv
$PRED
Y=THETA(1)+ETA(1)+ERR(1)
$THETA 1
$OMEGA 0.1
$SIGMA 1
$ESTIMATION METHOD=1"
read_model_from_string(s)
}
}
\seealso{
read_model : Read model from file

read_model_from_database : Read model from database
}
