% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{read_modelfit_results}
\alias{read_modelfit_results}
\title{read_modelfit_results}
\usage{
read_modelfit_results(path)
}
\arguments{
\item{path}{(str) Path to model file}
}
\value{
(ModelfitResults) Results object
}
\description{
Read results from external tool for a model
}
