\name{read_results}
\alias{read_results}
\title{read_results}
\description{
NA.

\bold{Usage}

\code{read_results(path)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.read_results}{API reference} (for correct rendering of equations, tables etc.).


This documentation was automatically generated from Pharmpy (0.26.0).}
