\name{remove_error}
\alias{remove_error}
\title{remove_error}
\description{
Remove error model..

\bold{Usage}

\code{remove_error(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.remove_error}{API reference} (for correct rendering of equations, tables etc.).
Remove error model.

This documentation was automatically generated from Pharmpy (0.25.1).}
\arguments{

\item{
model
}{
Model. Remove error model for this model
}
}
