% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_error_model}
\alias{remove_error_model}
\title{remove_error_model}
\usage{
remove_error_model(model)
}
\arguments{
\item{model}{(Model) Remove error model for this model}
}
\description{
Remove error model.
}
