% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_estimation_step}
\alias{remove_estimation_step}
\title{remove_estimation_step}
\usage{
remove_estimation_step(model, idx)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{idx}{(int) index of estimation step to remove}
}
\value{
model (Model)
}
\description{
Remove estimation step
}
