% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_estimation_step}
\alias{remove_estimation_step}
\title{remove_estimation_step}
\usage{
remove_estimation_step(model, idx)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{idx}{(numeric) index of estimation step to remove (starting from 0)}
}
\value{
(Model) Pharmpy model object
}
\description{
Remove estimation step
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- remove_estimation_step(model, 0)
ests <- model$estimation_steps
length(ests)
}
}
\seealso{
add_estimation_step

set_estimation_step

append_estimation_step_options

add_parameter_uncertainty_step

remove_parameter_uncertainty_step

set_evaluation_step
}
