% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_iiv}
\alias{remove_iiv}
\title{remove_iiv}
\usage{
remove_iiv(model, to_remove = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model to create block effect on.}

\item{to_remove}{(str, vector) Name/names of etas and/or name/names of individual parameters to remove.
If NULL, all etas that are IIVs will be removed. NULL is default.

Model
Reference to the same model}
}
\description{
Removes all IIV etas given a vector with eta names and/or parameter names.
}
\seealso{
remove_iov

add_iiv

add_iov
}
