\name{remove_iiv}
\alias{remove_iiv}
\title{remove_iiv}
\description{
Removes all IIV omegas given a list with eta names and/or parameter names..

\bold{Usage}

\code{remove_iiv(model, list_to_remove=None)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Removes all IIV omegas given a list with eta names and/or parameter names.

This documentation was automatically generated from Pharmpy (0.22.0).}
\arguments{

\item{
model
}{
Model, Pharmpy model to create block effect on.
}
\item{
list_to_remove
}{
str, list, Name/names of etas and/or parameter name to remove. If None, all etas that are IIVs will
}
}
