% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_iov}
\alias{remove_iov}
\title{remove_iov}
\usage{
remove_iov(model)
}
\arguments{
\item{model}{(Model) Pharmpy model to remove IOV from.

Model
Reference to the same model}
}
\description{
Removes all IOV etas
}
\seealso{
add_iiv

add_iov

remove_iiv
}
