\name{remove_iov}
\alias{remove_iov}
\title{remove_iov}
\description{
Removes all IOV omegas..

\bold{Usage}

\code{remove_iov(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.remove_iov}{API reference} (for correct rendering of equations, tables etc.).
Removes all IOV omegas.

This documentation was automatically generated from Pharmpy (0.25.1).}
\arguments{

\item{
model
}{
Model. Pharmpy model to remove IOV from.
}
}
