% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_iov}
\alias{remove_iov}
\title{remove_iov}
\usage{
remove_iov(model, to_remove = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model to remove IOV from.}

\item{to_remove}{(str, array(str) (optional)) Name/names of IOV etas to remove, e.g. 'ETA_IOV_1_1'.
If NULL, all etas that are IOVs will be removed. NULL is default.}
}
\value{
(Model) Reference to the same model
}
\description{
Removes all IOV etas given a vector with eta names.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
remove_iov(model)
}
}
\seealso{
add_iiv

add_iov

remove_iiv

add_pk_iiv
}
