% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_lag_time}
\alias{remove_lag_time}
\title{remove_lag_time}
\usage{
remove_lag_time(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Reference to same model
}
\description{
Remove lag time from the dose compartment of model.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
remove_lag_time(model)
}
}
\seealso{
set_transit_compartments

add_lag_time
}
