% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_lag_time}
\alias{remove_lag_time}
\title{remove_lag_time}
\usage{
remove_lag_time(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Remove lag time from the dose compartment of model.
}
