\name{remove_lag_time}
\alias{remove_lag_time}
\title{remove_lag_time}
\description{
Remove lag time from the dose compartment of model..

\bold{Usage}

\code{remove_lag_time(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.remove_lag_time}{API reference} (for correct rendering of equations, tables etc.).
Remove lag time from the dose compartment of model.

This documentation was automatically generated from Pharmpy (0.25.0).}
