% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_peripheral_compartment}
\alias{remove_peripheral_compartment}
\title{remove_peripheral_compartment}
\usage{
remove_peripheral_compartment(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Pharmpy model object
}
\description{
Remove a peripheral distribution compartment from model

Initial estimates:

==  ===================================================
n
==  ===================================================
2   (equation could not be rendered, see API doc on website)
3   (equation could not be rendered, see API doc on website)
==  ===================================================
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_peripheral_compartments(model, 2)
model <- remove_peripheral_compartment(model)
model$statements$ode_system
}
}
\seealso{
set_peripheral_compartment

add_peripheral_compartment
}
