\name{remove_peripheral_compartment}
\alias{remove_peripheral_compartment}
\title{remove_peripheral_compartment}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.remove_peripheral_compartment}{Link to Python API reference} (e.g. for correct rendering of equations)

Remove a peripheral distribution compartment from model

Initial estimates:

==  ===================================================

n

==  ===================================================

2   :math:`\mathsf{CL} = \mathsf{CL'}`,

:math:`\mathsf{QP1} = \mathsf{CL'}` and :math:`\mathsf{VP1} = \mathsf{VC'} \cdot 0.05`

3   :math:`\mathsf{QP1} = (\mathsf{QP1'} + \mathsf{QP2'}) / 2`,

:math:`\mathsf{VP1} = \mathsf{VP1'} + \mathsf{VP2'}`

==  ===================================================
}
\usage{
remove_peripheral_compartment(model)
}
\arguments{

}
\value{

}
