\name{remove_peripheral_compartment}
\alias{remove_peripheral_compartment}
\title{remove_peripheral_compartment}
\description{
Remove a peripheral distribution compartment from model.

\bold{Usage}

\code{remove_peripheral_compartment(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.remove_peripheral_compartment}{API reference} (for correct rendering of equations, tables etc.).
Remove a peripheral distribution compartment from model

Initial estimates:

==  ===================================================

n

==  ===================================================

2   :math:`{CL} = {CL'}`,

:math:`{QP1} = {CL'}` and :math:`{VP1} = {VC'} * 0.05`

3   :math:`{QP1} = ({QP1'} + {QP2'}) / 2`,

:math:`{VP1} = {VP1'} + {VP2'}`

==  ===================================================

This documentation was automatically generated from Pharmpy (0.26.1).}
