% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{remove_unused_parameters_and_rvs}
\alias{remove_unused_parameters_and_rvs}
\title{remove_unused_parameters_and_rvs}
\usage{
remove_unused_parameters_and_rvs(model)
}
\arguments{
\item{model}{(Model) Pharmpy model object}
}
\value{
(Model) Reference to same model object
}
\description{
Remove any parameters and rvs that are not used in the model statements
}
