% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{rename_symbols}
\alias{rename_symbols}
\title{rename_symbols}
\usage{
rename_symbols(model, new_names)
}
\arguments{
\item{model}{(Model) Pharmpy model object}

\item{new_names}{(list(str=str)) From old name or symbol to new name or symbol}
}
\value{
(Model) Pharmpy model object
}
\description{
Rename symbols in the model

Make sure that no name clash occur.
}
