% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{retrieve_model}
\alias{retrieve_model}
\title{retrieve_model}
\usage{
retrieve_model(source, name)
}
\arguments{
\item{source}{(str or Context) Source where to find models. Can be a path (as str or Path), or a
Context}

\item{name}{(str) Name of the model}
}
\value{
(Model) The model object
}
\description{
Retrieve a model from a context/tool run

Any models created and run by the tool can be
retrieved.
}
\examples{
\dontrun{
tooldir_path <- 'path/to/tool/directory'
model <- retrieve_model(tooldir_path, 'run1')
}

}
