% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{retrieve_modelfit_results}
\alias{retrieve_modelfit_results}
\title{retrieve_modelfit_results}
\usage{
retrieve_modelfit_results(context, name)
}
\arguments{
\item{context}{(Context) A previously opened context}

\item{name}{(str) Name of the model or a qualified name with a subcontext path, e.g.}
}
\value{
(ModelfitResults) The results object
}
\description{
Retrieve the modelfit results of a model
}
\examples{
\dontrun{
context <- open_context("iivsearch1")
results <- retrieve_modelfit_results(context, 'input')
}

}
