% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{retrieve_models}
\alias{retrieve_models}
\title{retrieve_models}
\usage{
retrieve_models(source, names = NULL)
}
\arguments{
\item{source}{(str or Context) Source where to find models. Can be a path (as str or Path), or a
Context}

\item{names}{(array(str) (optional)) List of names of the models to retrieve or NULL for all}
}
\value{
(vector) List of retrieved model objects
}
\description{
Retrieve models after a tool run

Any models created and run by the tool can be
retrieved.
}
\examples{
\dontrun{
tooldir_path <- 'path/to/tool/directory'
models <- retrieve_models(tooldir_path, names=c('run1'))
}

}
