% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_amd}
\alias{run_amd}
\title{run_amd}
\usage{
run_amd(model, mfl = NULL, lloq = NULL, order = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{mfl}{(str) MFL for search space for structural model}

\item{lloq}{(numeric) Lower limit of quantification. LOQ data will be removed.}

\item{order}{(vector) Runorder of components}
}
\value{
(Model) Reference to the same model object
}
\description{
Run Automatic Model Development (AMD) tool

Runs structural modelsearch, IIV building, and resmod
}
\examples{
\dontrun{
model <- load_example_model("pheno")
run_amd(model)
}
}
\seealso{
run_iiv

run_tool
}
