% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_amd}
\alias{run_amd}
\title{run_amd}
\usage{
run_amd(
  dataset_path,
  modeltype = "pk_oral",
  cl_init = 0.01,
  vc_init = 1,
  mat_init = 0.1,
  mfl = NULL,
  lloq = NULL,
  order = NULL,
  categorical = NULL,
  continuous = NULL
)
}
\arguments{
\item{dataset_path}{(Model) Path to a dataset}

\item{modeltype}{(str) Type of model to build. Either 'pk_oral' or 'pk_iv'}

\item{cl_init}{(numeric) Initial estimate for the population clearance}

\item{vc_init}{(numeric) Initial estimate for the central compartment population volume}

\item{mat_init}{(numeric) Initial estimate for the mean absorption time (not for iv models)}

\item{mfl}{(str) MFL for search space for structural model}

\item{lloq}{(numeric) Lower limit of quantification. LOQ data will be removed.}

\item{order}{(vector) Runorder of components}

\item{categorical}{(vector) List of categorical covariates}

\item{continuous}{(vector) List of continuouts covariates}
}
\value{
(Model) Reference to the same model object
}
\description{
Run Automatic Model Development (AMD) tool

Runs structural modelsearch, IIV building, and resmod
}
\examples{
\dontrun{
model <- load_example_model("pheno")
run_amd(model)
}
}
\seealso{
run_iiv

run_tool
}
