% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_covsearch}
\alias{run_covsearch}
\title{run_covsearch}
\usage{
run_covsearch(
  effects,
  p_forward = 0.05,
  p_backward = 0.01,
  max_steps = -1,
  algorithm = "scm-forward-then-backward",
  model = NULL
)
}
\arguments{
\item{effects}{(str | vector) The vector of candidate parameter-covariate effects to try, either as a
MFL sentence or in (optionally compact) tuple form.}

\item{p_forward}{(numeric) The p-value to use in the likelihood ratio test for forward steps}

\item{p_backward}{(numeric) The p-value to use in the likelihood ratio test for backward steps}

\item{max_steps}{(integer) The maximum number of search steps to make}

\item{algorithm}{(str) The search algorithm to use. Currently 'scm-forward' and
'scm-forward-then-backward' are supported.}

\item{model}{(Model) Pharmpy model}
}
\value{
(COVSearchResults) COVsearch tool result object
}
\description{
Run COVsearch tool. For more details, see :ref:\code{covsearch}.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
res <- run_covsearch([
}

}
