% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_iiv}
\alias{run_iiv}
\title{run_iiv}
\usage{
run_iiv(model, iiv_strategy = 0, rankfunc = "ofv", cutoff = NULL, path = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{iiv_strategy}{(integer) How IIVs should be added to start model. Default is 0 (no added IIVs)}

\item{rankfunc}{(str) Which ranking function should be used (OFV, AIC, BIC). Default is OFV}

\item{cutoff}{(numeric) Cutoff for which value of the ranking function that is considered significant. Default
is 3.84}

\item{path}{(Path) Path of rundirectory}
}
\value{
(Model) Reference to the same model object
}
\description{
Run IIV tool

Runs two IIV workflows: testing the number of etas and testing which block structure
}
\examples{
\dontrun{
model <- load_example_model("pheno")
run_iiv(model)
}
}
\seealso{
run_amd

run_tool
}
