% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_iiv}
\alias{run_iiv}
\title{run_iiv}
\usage{
run_iiv(
  model,
  add_iivs = FALSE,
  iiv_as_fullblock = FALSE,
  rankfunc = "ofv",
  cutoff = NULL,
  path = NULL
)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{add_iivs}{(logical) Whether to add IIV on structural parameters. Default is FALSE}

\item{iiv_as_fullblock}{(logical) Whether added etas should be as a fullblock. Default is FALSE}

\item{rankfunc}{(str) Which ranking function should be used (OFV, AIC, BIC). Default is OFV}

\item{cutoff}{(numeric) Cutoff for which value of the ranking function that is considered significant. Default
is 3.84}

\item{path}{(Path) Path of rundirectory}
}
\value{
(Model) Reference to the same model object
}
\description{
Run IIV tool

Runs two IIV workflows: testing the number of etas and testing which block structure
}
\examples{
\dontrun{
model <- load_example_model("pheno")
run_iiv(model)
}
}
\seealso{
run_amd

run_tool
}
