% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_iivsearch}
\alias{run_iivsearch}
\title{run_iivsearch}
\usage{
run_iivsearch(
  algorithm,
  iiv_strategy = "no_add",
  rankfunc = "bic",
  cutoff = NULL,
  model = NULL
)
}
\arguments{
\item{algorithm}{(str) Which algorithm to run (brute_force, brute_force_no_of_etas, brute_force_block_structure)}

\item{iiv_strategy}{(str) If/how IIV should be added to start model. Possible strategies are 'no_add', 'diagonal', or
'fullblock'. Default is 'no_add'}

\item{rankfunc}{(str) Which ranking function should be used (OFV, AIC, BIC). Default is BIC}

\item{cutoff}{(numeric) Cutoff for which value of the ranking function that is considered significant. Default
is NULL (all models will be ranked)}

\item{model}{(Model) Pharmpy model}
}
\value{
(IIVResults) IIVsearch tool result object
}
\description{
Run IIVsearch tool. For more details, see :ref:\code{iivsearch}.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
run_iivsearch('brute_force', model=model)
}

}
