% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_iivsearch}
\alias{run_iivsearch}
\title{run_iivsearch}
\usage{
run_iivsearch(
  algorithm,
  iiv_strategy = "no_add",
  rank_type = "mbic",
  cutoff = NULL,
  results = NULL,
  model = NULL,
  keep = NULL,
  strictness = "minimization_successful or (rounding_errors and sigdigs>=0.1)",
  ...
)
}
\arguments{
\item{algorithm}{(str) Which algorithm to run.}

\item{iiv_strategy}{(str) If/how IIV should be added to start model. Default is 'no_add'.}

\item{rank_type}{(str) Which ranking type should be used. Default is mBIC.}

\item{cutoff}{(numeric (optional)) Cutoff for which value of the ranking function that is considered significant. Default
is NULL (all models will be ranked)}

\item{results}{(ModelfitResults (optional)) Results for model}

\item{model}{(Model (optional)) Pharmpy model}

\item{keep}{(array(str) (optional)) List of IIVs to keep}

\item{strictness}{(str (optional)) Strictness criteri}

\item{...}{Arguments to pass to tool}
}
\value{
(IIVSearchResults) IIVsearch tool result object
}
\description{
Run IIVsearch tool. For more details, see :ref:\code{iivsearch}.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
results <- load_example_modelfit_results("pheno")
run_iivsearch('brute_force', results=results, model=model)
}

}
