% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_iovsearch}
\alias{run_iovsearch}
\title{run_iovsearch}
\usage{
run_iovsearch(
  column = "OCC",
  list_of_parameters = NULL,
  rank_type = "bic",
  cutoff = NULL,
  distribution = "same-as-iiv",
  results = NULL,
  model = NULL,
  strictness = "minimization_successful or (rounding_errors and sigdigs>=0.1)",
  ...
)
}
\arguments{
\item{column}{(str) Name of column in dataset to use as occasion column (default is 'OCC')}

\item{list_of_parameters}{(array(str) (optional)) List of parameters to test IOV on, if none all parameters with IIV will be tested (default)}

\item{rank_type}{(str) Which ranking type should be used (OFV, AIC, BIC). Default is BIC}

\item{cutoff}{(numeric (optional)) Cutoff for which value of the ranking type that is considered significant. Default
is NULL (all models will be ranked)}

\item{distribution}{(str) Which distribution added IOVs should have (default is same-as-iiv)}

\item{results}{(ModelfitResults (optional)) Results for model}

\item{model}{(Model (optional)) Pharmpy model}

\item{strictness}{(str (optional)) Strictness criteri}

\item{...}{Arguments to pass to tool}
}
\value{
(IOVSearchResults) IOVSearch tool result object
}
\description{
Run IOVsearch tool. For more details, see :ref:\code{iovsearch}.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
results <- load_example_modelfit_results("pheno")
run_iovsearch('OCC', results=results, model=model)
}

}
