% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_linearize}
\alias{run_linearize}
\title{run_linearize}
\usage{
run_linearize(model = NULL, model_name = "linbase", description = "", ...)
}
\arguments{
\item{model}{(Model (optional)) Pharmpy model.}

\item{model_name}{(str) New name of linearized model. The default is "linbase".}

\item{description}{(str) Description of linaerized model. The default is ""}

\item{...}{Arguments to pass to tool}
}
\value{
(LinearizeResults) Linaerize tool results object.
}
\description{
Run linaerization procedure
}
