% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_pdsearch}
\alias{run_pdsearch}
\title{run_pdsearch}
\usage{
run_pdsearch(
  dataset,
  treatment_variable,
  strictness = "minimization_successful or (rounding_errors and sigdigs>=0.1)",
  parameter_uncertainty_method = NULL,
  ...
)
}
\arguments{
\item{dataset}{(str) A PD dataset}

\item{treatment_variable}{(str) Name of the variable representing the treatment, e.g. TRT, DOSE or AUC}

\item{strictness}{(str) Strictness criteria}

\item{parameter_uncertainty_method}{(str (optional)) Parameter uncertainty method. Will be used in ranking models if strictness includes
parameter uncertaint}

\item{...}{Arguments to pass to tool}
}
\value{
(PDSearchResults) PDSearch tool results object.
}
\description{
Build a PD model
}
