% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_pdsearch}
\alias{run_pdsearch}
\title{run_pdsearch}
\usage{
run_pdsearch(
  input,
  type,
  treatment_variable = NULL,
  kpd_driver = "ir",
  results = NULL,
  strictness = "minimization_successful or (rounding_errors and sigdigs>=0.1)",
  parameter_uncertainty_method = NULL,
  ...
)
}
\arguments{
\item{input}{(str or Model) A PD/KPD dataset or PD/KPD model}

\item{type}{(str) Type of PD model to build ('pd' or 'kpd')}

\item{treatment_variable}{(str (optional)) Name of the variable representing the treatment, e.g. TRT, DOSE or AUC. Do not use if \code{type} is 'kpd'}

\item{kpd_driver}{(str) Driver for KPD model (virtual infusion rate 'ir' or 'amount')}

\item{results}{(ModelfitResults (optional)) Results to input model}

\item{strictness}{(str) Strictness criteria}

\item{parameter_uncertainty_method}{(str (optional)) Parameter uncertainty method. Will be used in ranking models if strictness includes
parameter uncertaint}

\item{...}{Arguments to pass to tool}
}
\value{
(PDSearchResults) PDSearch tool results object.
}
\description{
Build a PD model
}
