% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_qa}
\alias{run_qa}
\title{run_qa}
\usage{
run_qa(model = NULL, results = NULL, linearize = FALSE, skip = NULL, ...)
}
\arguments{
\item{model}{(Model (optional)) Pharmpy model}

\item{results}{(ModelfitResults (optional)) Results of model}

\item{linearize}{(logical) Whether or not to use linearization when running the tool}

\item{skip}{(array(str) (optional)) A vector of sections to ski}

\item{...}{Arguments to pass to tool}
}
\value{
(QAResults) QA tool result object
}
\description{
Run QA tool.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
results <- load_example_modelfit_results("pheno")
run_qa(model=model, results=results, linearize=FALSE, skip=c('fullblock'))
}

}
