% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_structsearch}
\alias{run_structsearch}
\title{run_structsearch}
\usage{
run_structsearch(route, type, results = NULL, model = NULL, ...)
}
\arguments{
\item{route}{(str) Route of administration. Either 'pk' or 'oral'}

\item{type}{(str) Type of model. Currently only 'tmdd'}

\item{results}{(ModelfitResults (optional)) Results for the start model}

\item{model}{(Model (optional)) Pharmpy start mode}

\item{...}{Arguments to pass to tool}
}
\value{
(StructSearchResult) structsearch tool result object
}
\description{
Run the structsearch tool. For more details, see :ref:\code{structsearch}.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
results <- load_example_modelfit_results("pheno")
run_structsearch(model_type='tmdd', results=results, model=model)
}

}
