% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{run_tool}
\alias{run_tool}
\title{run_tool}
\usage{
run_tool(name, ...)
}
\arguments{
\item{name}{(str) Name of tool to run}

\item{...}{Arguments to pass to tool}
}
\value{
(Results object for tool)
}
\description{
Run tool workflow
}
\examples{
\dontrun{
model <- load_example_model("pheno")
res <- run_tool("resmod", model)
}

}
