\name{seq_zo_fo_absorption}
\alias{seq_zo_fo_absorption}
\title{seq_zo_fo_absorption}
\description{
Set or change to sequential zero order first order absorption rate.

\bold{Usage}

\code{seq_zo_fo_absorption(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Set or change to sequential zero order first order absorption rate. Initial estimate for

absorption rate is set the previous rate if available, otherwise it is set to the time of

first observation/2 is used.
}
\arguments{

\item{
model
}{
Model, Model to set or change absorption rate
}
}
