% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_bolus_absorption}
\alias{set_bolus_absorption}
\title{set_bolus_absorption}
\usage{
set_bolus_absorption(model)
}
\arguments{
\item{model}{(Model) Model to set or change absorption rate}
}
\value{
(Model) Reference to same model
}
\description{
Set or change to bolus absorption rate.

Currently lagtime together with bolus absorption is not supported.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
set_bolus_absorption(model)
model$statements$ode_system
}
}
\seealso{
set_zero_order_absorption

set_first_order_absorption
}
