% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_covariates}
\alias{set_covariates}
\title{set_covariates}
\usage{
set_covariates(model, covariates)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{covariates}{(vector) List of column names}
}
\value{
(Model) Reference to the same Pharmpy model object
}
\description{
Set columns in the dataset to be covariates in the datainfo
}
