% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_dataset}
\alias{set_dataset}
\title{set_dataset}
\usage{
set_dataset(model, path_or_df, datatype = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{path_or_df}{(str or data.frame) Dataset path or dataframe}

\item{datatype}{(str (optional)) Type of dataset (optional)}
}
\value{
(Model) Pharmpy model with new dataset and updated datainfo
}
\description{
Load the dataset given datainfo
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- unload_dataset(model)
dataset_path <- model$datainfo$path
model$dataset is NULL
model <- set_dataset(model, dataset_path, datatype='nonmem')
model$dataset
}

}
