% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_direct_effect}
\alias{set_direct_effect}
\title{set_direct_effect}
\usage{
set_direct_effect(model, expr, variable = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{expr}{(str) Name of PD effect function.}

\item{variable}{(str (optional)) Name of variable to use (if NULL concentration will be used)}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Add an effect to a model.

Effects are by default using concentratrion, but any user specified
variable in the model can be used. Implemented PD models are:
\itemize{
\item Linear:
}

(equation could not be rendered, see API doc on website)
\itemize{
\item Emax:
}

(equation could not be rendered, see API doc on website)
\itemize{
\item Step effect:
}

(equation could not be rendered, see API doc on website)
\itemize{
\item Sigmoidal:
}

(equation could not be rendered, see API doc on website)
\itemize{
\item Log-linear:
}

(equation could not be rendered, see API doc on website)

(equation could not be rendered, see API doc on website)
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_direct_effect(model, "linear")
model$statements$find_assignment("E")
}

}
