% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_direct_effect}
\alias{set_direct_effect}
\title{set_direct_effect}
\usage{
set_direct_effect(model, expr)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{expr}{(str) Name of PD effect function. Valid names are: baseline, linear, emax, sigmoid, step and loglin}
}
\value{
(Model) Pharmpy model object
}
\description{
Add an effect to a model.

Implemented PD models are:
\itemize{
\item Baseline:
(equation could not be rendered, see API doc on website)
\item Linear:
(equation could not be rendered, see API doc on website)
\item Emax:
(equation could not be rendered, see API doc on website)
\item Step effect:
(equation could not be rendered, see API doc on website)
\item Sigmoidal:
(equation could not be rendered, see API doc on website)
\item Log-linear:
(equation could not be rendered, see API doc on website)
}

(equation could not be rendered, see API doc on website)
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_direct_effect(model, "linear")
model$statements$find_assignment("E")
}

}
