% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_dtbs_error_model}
\alias{set_dtbs_error_model}
\title{set_dtbs_error_model}
\usage{
set_dtbs_error_model(model, fix_to_log = FALSE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{fix_to_log}{(logical) Set to TRUE to fix lambda and zeta to 0, i.e. emulating log-transformed data}
}
\value{
(Model) Pharmpy model object
}
\description{
Dynamic transform both sides
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_dtbs_error_model(model)
}

}
