% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_estimation_step}
\alias{set_estimation_step}
\title{set_estimation_step}
\usage{
set_estimation_step(
  model,
  method,
  interaction = TRUE,
  options = list(),
  est_idx = 0
)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{method}{(str) estimation method to change to}

\item{interaction}{(logical) whether to use interaction or not, default is true}

\item{options}{(list) any additional options. Note that this removes old options}

\item{est_idx}{(int) index of estimation step, default is 0 (first estimation step)

Model
Reference to the same model object}
}
\description{
Set estimation step

Sets estimation step for a model. Methods currently supported are:
FO, FOCE, ITS, LAPLACE, IMPMAP, IMP, SAEM
}
\seealso{
add_estimation_step

remove_estimation_step
}
