% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_first_order_absorption}
\alias{set_first_order_absorption}
\title{set_first_order_absorption}
\usage{
set_first_order_absorption(model)
}
\arguments{
\item{model}{(Model) Model to set or change to use first order absorption rate}
}
\value{
(Model) Pharmpy model object
}
\description{
Set or change to first order absorption rate.

Initial estimate for absorption rate is set to
the previous rate if available, otherwise it is set to the time of first observation/2.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_first_order_absorption(model)
model$statements$ode_system
}
}
\seealso{
set_bolus_order_absorption

set_zero_order_absorption
}
