\name{set_initial_estimates}
\alias{set_initial_estimates}
\title{set_initial_estimates}
\description{
Set initial estimates.

\bold{Usage}

\code{set_initial_estimates(model, inits)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Set initial estimates

This documentation was automatically generated from Pharmpy (0.22.0).}
\arguments{

\item{
model
}{
Model, inits : A dictionary of parameter: init for parameters to change
}
\item{
inits
}{
A dictionary of parameter, NULL
}
}
\value{
model : Model
}
