% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_initial_estimates}
\alias{set_initial_estimates}
\title{set_initial_estimates}
\usage{
set_initial_estimates(model, inits)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{inits}{(list(str=numeric)) A list of parameter init for parameters to change}
}
\value{
(Model) Pharmpy model object
}
\description{
Set initial estimates
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_initial_estimates(model, {'PTVCL': 2.0})
model$parameters['PTVCL']
}
}
\seealso{
fix_parameters_to : Fixing and setting parameter initial estimates in the same function

unfix_paramaters_to : Unfixing parameters and setting a new initial estimate in the same
}
