\name{set_initial_estimates}
\alias{set_initial_estimates}
\title{set_initial_estimates}
\description{
Set initial estimates.

\bold{Usage}

\code{set_initial_estimates(model, inits)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.set_initial_estimates}{API reference} (for correct rendering of equations, tables etc.).
Set initial estimates

This documentation was automatically generated from Pharmpy (0.26.0).}
\arguments{

\item{
model
}{
Model. Pharmpy model
}
\item{
inits
}{
list. A list of parameter init for parameters to change
}
}
\value{
model : Model
}
