% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_n_transit_compartments}
\alias{set_n_transit_compartments}
\title{set_n_transit_compartments}
\usage{
set_n_transit_compartments(model, keep_depot = TRUE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{keep_depot}{(logical) FALSE to convert depot compartment into a transit compartment}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Set the n-transit compartments model

This is the absorption delay model where the number of transit compartments is a parameter
to be estimated. Initial estimate for absorption rate is
set the previous rate if available, otherwise it is set to the time of first observation/2.
Initial estimate for the number of transit compartments is set to 2.

Currently only handles a single oral route of administration.
Assumes complete absorption between doses
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_n_transit_compartments(model)
}
}
\seealso{
add_lag_time

set_transit_compartments
}
