\name{set_name}
\alias{set_name}
\title{set_name}
\description{
NULL.

\bold{Usage}

\code{set_name(model, name)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.set_name}{API reference} (for correct rendering of equations, tables etc.).


This documentation was automatically generated from Pharmpy (0.22.0).}
