% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_name}
\alias{set_name}
\title{set_name}
\usage{
set_name(model, new_name)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{new_name}{(str) New name of model}
}
\value{
(Model) Reference to the same model object
}
\description{
Set name of model object
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$name
set_name(model, "run2")
model$name
}

}
