\name{set_ode_solver}
\alias{set_ode_solver}
\title{set_ode_solver}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.set_ode_solver}{Link to Python API reference} (e.g. for correct rendering of equations)


}
\usage{
set_ode_solver(model, solver)
}
\arguments{

}
\value{

}
