% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_ode_solver}
\alias{set_ode_solver}
\title{set_ode_solver}
\usage{
set_ode_solver(model, solver)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{solver}{(str) Solver to use}
}
\description{
Sets ode solver
}
