\name{set_ode_solver}
\alias{set_ode_solver}
\title{set_ode_solver}
\description{
NA.

\bold{Usage}

\code{set_ode_solver(model, solver)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.set_ode_solver}{API reference} (for correct rendering of equations, tables etc.).


This documentation was automatically generated from Pharmpy (0.22.10).}
