\name{set_peripheral_compartments}
\alias{set_peripheral_compartments}
\title{set_peripheral_compartments}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.set_peripheral_compartments}{Link to Python API reference} (e.g. for correct rendering of equations)


}
\usage{
set_peripheral_compartments(model, n)
}
\arguments{

}
\value{

}
