\name{set_peripheral_compartments}
\alias{set_peripheral_compartments}
\title{set_peripheral_compartments}
\description{
NULL.

\bold{Usage}

\code{set_peripheral_compartments(model, n)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.set_peripheral_compartments}{API reference} (for correct rendering of equations, tables etc.).


This documentation was automatically generated from Pharmpy (0.22.0).}
