% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_peripheral_compartments}
\alias{set_peripheral_compartments}
\title{set_peripheral_compartments}
\usage{
set_peripheral_compartments(model, n)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{n}{(numeric) Number of transit compartments}
}
\value{
(Model) Pharmpy model object
}
\description{
Sets the number of peripheral compartments to a specified number.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_peripheral_compartments(model, 2)
model$statements$ode_system
}
}
\seealso{
add_peripheral_compartment

remove_peripheral_compartment
}
