% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_property}
\alias{set_property}
\title{set_property}
\usage{
set_property(model_or_datainfo, column, property, value)
}
\arguments{
\item{model_or_datainfo}{(Model or DataInfo) Model object or DataInfo object}

\item{column}{(str) Name of a column. If the column contains multiple variables, e.g. DV
with multiple DVs, the ID can be specified with a colon. For example "DV:1"
will mean the DV column only when DVID is 1.}

\item{property}{(str) Name of the property to set}

\item{value}{(any) Value of the property to set}
}
\value{
(Model | DataInfo) An updated Model or DataInfo object
}
\description{
Specify a property of a column

See :py:attr:\code{pharmpy.DataInfo.properties} for documentation on data properties.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_property(model, "APGR", "categories", c(0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10))
}
}
\seealso{
set_unit - Set unit of a data variable
}
