% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_tmdd}
\alias{set_tmdd}
\title{set_tmdd}
\usage{
set_tmdd(model, type)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{type}{(str) Type of TMDD model}
}
\value{
(Model) Pharmpy model object
}
\description{
Sets target mediated drug disposition

Sets target mediated drug disposition to a PK model.

Supported models are full, ib, cr, crib, qss, wagner and mmapp.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_tmdd(model, "full")
}

}
