\name{set_transit_compartments}
\alias{set_transit_compartments}
\title{set_transit_compartments}
\description{
Set the number of transit compartments of model.

\bold{Usage}

\code{set_transit_compartments(model, n)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.set_transit_compartments}{API reference} (for correct rendering of equations, tables etc.).
Set the number of transit compartments of model. Initial estimate for absorption rate is

set the previous rate if available, otherwise it is set to the time of first observation/2

is used.

This documentation was automatically generated from Pharmpy (0.26.0).}
