% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_transit_compartments}
\alias{set_transit_compartments}
\title{set_transit_compartments}
\usage{
set_transit_compartments(model, n)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{n}{(int) Number of transit compartments}
}
\description{
Set the number of transit compartments of model. Initial estimate for absorption rate is
set the previous rate if available, otherwise it is set to the time of first observation/2
is used.
}
