% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_upper_bounds}
\alias{set_upper_bounds}
\title{set_upper_bounds}
\usage{
set_upper_bounds(model, bounds, strict = TRUE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{bounds}{(list(str=numeric)) A list of parameter bounds for parameters to change}

\item{strict}{(logical) Whether all parameters in input need to exist in the model. Default is TRUE}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Set parameter upper bounds
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_upper_bounds(model, list('POP_CL'=10))
model$parameters['POP_CL']
}
}
\seealso{
set_lower_bounds : Set parameter lower bounds

unconstrain_parameters : Remove all constraints of parameters
}
