% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_weibull_absorption}
\alias{set_weibull_absorption}
\title{set_weibull_absorption}
\usage{
set_weibull_absorption(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Set or change to Weibull type absorption

The Weibull absorption has an abosption rate varying with time (or rather time after dose).
(equation could not be rendered, see API doc on website)

Initial parameter estimates will be set differently depending on whether the original model
has MAT and/or MDT.

=== === ========= =======================================================
MAT MDT (equation could not be rendered, see API doc on website)
=== === ========= =======================================================
yes yes 1.5       (equation could not be rendered, see API doc on website)
yes no  1.0       MAT
no  yes 1.0       MDT
no  no  1.5       Same as if having MAT, but use min observation time * 2
=== === ========= =======================================================

If multiple doses are fed into the affected compartment, currently only iv+oral
doses (one of each) is supported.

Weibull absorption cannot be used together with lag time or transit compartments.

Assumes that absorption of one dose is complete when the next dose is given.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- set_weibull_absorption(model)
model$statements$ode_system
}
}
\seealso{
set_zero_order_absorption

set_first_order_absorption
}
