\name{set_weighted_error_model}
\alias{set_weighted_error_model}
\title{set_weighted_error_model}
\description{
Encode error model with one epsilon and W as weight.

\bold{Usage}

\code{set_weighted_error_model(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.set_weighted_error_model}{API reference} (for correct rendering of equations, tables etc.).
Encode error model with one epsilon and W as weight

This documentation was automatically generated from Pharmpy (0.25.1).}
