% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{set_zero_order_elimination}
\alias{set_zero_order_elimination}
\title{set_zero_order_elimination}
\usage{
set_zero_order_elimination(model)
}
\arguments{
\item{model}{(Model) Pharmpy model}
}
\description{
Sets elimination to zero order. Initial estimate for KM is set to 1\% of smallest
observation.
}
