\name{split_rv_block}
\alias{split_rv_block}
\title{split_rv_block}
\description{
Splits an block structure given a list of etas to separate..

\bold{Usage}

\code{split_rv_block(model, list_of_rvs=None)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Splits an block structure given a list of etas to separate.

This documentation was automatically generated from Pharmpy (0.22.0).}
\arguments{

\item{
model
}{
Model, Pharmpy model to create block effect on.
}
\item{
list_of_rvs
}{
str, list, Name/names of etas to split from block structure. If None, all etas that are IIVs and
}
}
