% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{summarize_errors}
\alias{summarize_errors}
\title{summarize_errors}
\usage{
summarize_errors(models)
}
\arguments{
\item{models}{(vector, Model) List of models or single model}
}
\value{
(data.frame) A DataFrame of errors with model name, category (error or warning), and an integer as index, an empty DataFrame if there were no errors or warnings found.
}
\description{
Summarize errors and warnings from one or multiple model runs.

Summarize the errors and warnings found after running the model/models.
}
\examples{
\dontrun{
model <- load_example_model("pheno")
summarize_errors(model)
}

}
