% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{summarize_modelfit_results}
\alias{summarize_modelfit_results}
\title{summarize_modelfit_results}
\usage{
summarize_modelfit_results(models)
}
\arguments{
\item{models}{(vector) List of models}
}
\description{
Summarize results of multiple model runs, includes runtime, ofv and parameter estimates
(with errors).
}
