% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{summarize_modelfit_results}
\alias{summarize_modelfit_results}
\title{summarize_modelfit_results}
\usage{
summarize_modelfit_results(models)
}
\arguments{
\item{models}{(vector, Model) List of models or single model}
}
\value{
(data.frame) A DataFrame of modelfit results, one row per model.
}
\description{
Summarize results of model runs

Summarize different results after fitting a model, includes runtime, ofv,
and parameter estimates (with errors).
}
\examples{
\dontrun{
model <- load_example_model("pheno")
summarize_modelfit_results(c(model))
}

}
