% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{transform_etas_boxcox}
\alias{transform_etas_boxcox}
\title{transform_etas_boxcox}
\usage{
transform_etas_boxcox(model, list_of_etas = NULL)
}
\arguments{
\item{model}{(Model) Pharmpy model to apply boxcox transformation to.}

\item{list_of_etas}{(str, vector) Name/names of etas to transform. If NULL, all etas will be transformed (default).}
}
\description{
Applies a boxcox transformation to specified etas from a :class:\code{pharmpy.model}. Initial
estimate for lambda is 0.1 with bounds (-3, 3).
}
