% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{unconstrain_parameters}
\alias{unconstrain_parameters}
\title{unconstrain_parameters}
\usage{
unconstrain_parameters(model, parameter_names)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameter_names}{(array(str)) Remove all constraints for the listed parameters}
}
\value{
(Model) Pharmpy model object
}
\description{
Remove all constraints from parameters
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model$parameters['PTVCL']
model <- unconstrain_parameters(model, c('PTVCL'))
model$parameters['PTVCL']
}
}
\seealso{
set_lower_bounds : Set parameter lower bounds

set_upper_bounds : Set parameter upper bounds

unfix_parameters : Unfix parameters
}
