% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{unfix_parameters}
\alias{unfix_parameters}
\title{unfix_parameters}
\usage{
unfix_parameters(model, parameter_names)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameter_names}{(vector or str) one parameter name or a vector of parameter names}
}
\value{
model (Model)
}
\description{
Unfix parameters

Unfix all listed parameters
}
