% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{unfix_parameters}
\alias{unfix_parameters}
\title{unfix_parameters}
\usage{
unfix_parameters(model, parameter_names)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameter_names}{(vector or str) one parameter name or a vector of parameter names}
}
\value{
Model (Reference to the same model object)
}
\description{
Unfix parameters

Unfix all listed parameters
}
\seealso{
unfix_paramaters_to : Unfixing parameters and setting a new initial estimate in the same

function

fix_parameters : Fix parameters

fix_parameters_to : Fixing and setting parameter initial estimates in the same function
}
