\name{unfix_parameters}
\alias{unfix_parameters}
\title{unfix_parameters}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.unfix_parameters}{Link to Python API reference} (e.g. for correct rendering of equations)

Unfix parameters

Unfix all listed parameters
}
\usage{
unfix_parameters(model, parameter_names)
}
\arguments{

}
\value{
model: Model
}
