\name{unfix_parameters}
\alias{unfix_parameters}
\title{unfix_parameters}
\description{
Unfix parameters.

\bold{Usage}

\code{unfix_parameters(model, parameter_names)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.unfix_parameters}{API reference} (for correct rendering of equations, tables etc.).
Unfix parameters

Unfix all listed parameters

This documentation was automatically generated from Pharmpy (0.25.1).}
\arguments{

\item{
model
}{
Model. Pharmpy model
}
\item{
parameter_names
}{
vector or character. one parameter name or a vector of parameter names
}
}
\value{
model : Model
}
