% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{unfix_parameters}
\alias{unfix_parameters}
\title{unfix_parameters}
\usage{
unfix_parameters(model, parameter_names)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameter_names}{(array(str) or str) one parameter name or a vector of parameter names}
}
\value{
(Model) Pharmpy model object
}
\description{
Unfix parameters

Unfix all listed parameters
}
\examples{
\dontrun{
model <- load_example_model("pheno")
model <- fix_parameters(model, c('POP_CL', 'POP_VC'))
model$parameters$fix
model <- unfix_parameters(model, 'POP_CL')
model$parameters$fix
}
}
\seealso{
unfix_paramaters_to : Unfixing parameters and setting a new initial estimate in the same

function

fix_parameters : Fix parameters

fix_or_unfix_parameters : Fix or unfix parameters (given boolean)

fix_parameters_to : Fixing and setting parameter initial estimates in the same function

unconstrain_parameters : Remove all constraints of parameters
}
