% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{unfix_parameters_to}
\alias{unfix_parameters_to}
\title{unfix_parameters_to}
\usage{
unfix_parameters_to(model, parameter_names, values)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{parameter_names}{(vector or str) one parameter name or a vector of parameter names}

\item{values}{(vector or float) one value or a vector of values (must be equal to number of parameter_names)

Model
Reference to the same model object}
}
\description{
Unix parameters to

Unfix all listed parameters to specified value/values
}
