\name{unfix_parameters_to}
\alias{unfix_parameters_to}
\title{unfix_parameters_to}
\description{
Unix parameters to.

\bold{Usage}

\code{unfix_parameters_to(model, parameter_names, values)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Unix parameters to

Unfix all listed parameters to specified value/values

This documentation was automatically generated from Pharmpy (0.22.0).}
\arguments{

\item{
model
}{
Model, parameter_names : list or str
}
\item{
parameter_names
}{
list or str, one parameter name or a list of parameter names
}
\item{
values
}{
list or int, one value or a list of values (must be equal to number of parameter_names)
}
}
\value{
model : Model
}
