\name{update_inits}
\alias{update_inits}
\title{update_inits}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.update_inits}{Link to Python API reference} (e.g. for correct rendering of equations)

Updates initial estimates from previous output. Can be forced if no initial

individual estimates have been read.
}
\usage{
update_inits(model, force_individual_estimates=False)
}
\arguments{

\item{
model
}{
Model, Pharmpy model to create block effect on.
}
\item{
force_individual_estimates
}{
bool, Whether update of initial individual estimates should be forced.
}
}
\value{

}
