\name{update_inits}
\alias{update_inits}
\title{update_inits}
\description{
Updates initial estimates from previous output.

\bold{Usage}

\code{update_inits(model, force_individual_estimates=FALSE)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.update_inits}{API reference} (for correct rendering of equations, tables etc.).
Updates initial estimates from previous output. Can be forced if no initial

individual estimates have been read.

This documentation was automatically generated from Pharmpy (0.22.0).}
\arguments{

\item{
model
}{
Model. Pharmpy model to create block effect on.
}
\item{
force_individual_estimates
}{
bool. Whether update of initial individual estimates should be forced.
}
}
