\name{update_source}
\alias{update_source}
\title{update_source}
\description{
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html?highlight=add_cov#pharmpy.modeling.update_source}{Link to Python API reference} (e.g. for correct rendering of equations)

Update source

Let the code of the underlying source language be updated to reflect

changes in the model object.
}
\usage{
update_source(model)
}
\arguments{

}
\value{

}
