\name{update_source}
\alias{update_source}
\title{update_source}
\description{
Update source.

\bold{Usage}

\code{update_source(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.update_source}{API reference} (for correct rendering of equations, tables etc.).
Update source

Let the code of the underlying source language be updated to reflect

changes in the model object.

This documentation was automatically generated from Pharmpy (0.23.4).}
