% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{write_model}
\alias{write_model}
\title{write_model}
\usage{
write_model(model, path = "", force = TRUE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{path}{(str) Destination path}

\item{force}{(logical) Force overwrite, default is TRUE}
}
\value{
(Model) Updated Pharmpy model
}
\description{
Write model code to file

An updated Pharmpy model is returned. This will have a new name based on the filename
and it might have updates to the model code (e.g. for $DATA in NONMEM models).
}
\examples{
\dontrun{
model <- load_example_model("pheno")
write_model(model)
}

}
