% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{write_model}
\alias{write_model}
\title{write_model}
\usage{
write_model(model, path = "", force = TRUE)
}
\arguments{
\item{model}{(Model) Pharmpy model}

\item{path}{(str) Destination path}

\item{force}{(logical) Force overwrite, default is TRUE

Model
Reference to the same model object}
}
\description{
Write model code to file
}
