\name{write_model}
\alias{write_model}
\title{write_model}
\description{
Write model to file.

\bold{Usage}

\code{write_model(model, path='', force=FALSE)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html#pharmpy.modeling.write_model}{API reference} (for correct rendering of equations, tables etc.).
Write model to file

This documentation was automatically generated from Pharmpy (0.22.10).}
