% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_wrapper.R
\name{write_results}
\alias{write_results}
\title{write_results}
\usage{
write_results(results, path, lzma = FALSE, csv = FALSE)
}
\arguments{
\item{results}{(Results) Pharmpy results object}

\item{path}{(str) Path to results file}

\item{lzma}{(logical) TRUE for lzma compression. Not applicable to csv file}

\item{csv}{(logical) Save as csv file}
}
\description{
Write results object to json (or csv) file

Note that the csv-file cannot be read into a results object again.
}
