\name{zero_order_elimination}
\alias{zero_order_elimination}
\title{zero_order_elimination}
\description{
Sets elimination to zero order.

\bold{Usage}

\code{zero_order_elimination(model)}

}
\details{
Link to Python
\href{https://pharmpy.github.io/latest/reference/pharmpy.modeling.html}{API reference} (for correct rendering of equations, tables etc.).
Sets elimination to zero order. Initial estimate for KM is set to 1% of smallest

observation.
}
